% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{activate_provenance}
\alias{activate_provenance}
\title{Activate the provenance in the edibble design object}
\usage{
activate_provenance(
  .edibble,
  overwrite = c("graph", "anatomy", "recipe", "validation", "simulate")
)
}
\arguments{
\item{.edibble}{An edibble object.}

\item{overwrite}{What object to overwrite in the provenance object.}
}
\value{
A Provenance object.
}
\description{
This is a developer function to create a new Kitchen class with
the existing design.
}
\examples{
activate_provenance(takeout())
}
