% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{add_context}
\alias{add_context}
\title{Describe context related to experiment}
\usage{
add_context(.data, ..., .overwrite = TRUE)
}
\arguments{
\item{.data}{An edibble design.}

\item{...}{Name-value pairs where name is the context name and value is context
in string. Unnamed values will appended as points at the end.}

\item{.overwrite}{If the named context already exissts, it will overwrite.
By default it is TRUE.}
}
\description{
This is a function that stores a simple context about the experiment.
If the context already exists then it will be overwritten or ignored.
}
\examples{
start_design("COVID-19") \%>\%
  add_context(question = "Does Pfizer vaccine work?",
                 where = "Tested in lab")
}
\seealso{
\code{\link[=sort_context]{sort_context()}} for sorting context alphabetically.
}
