% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{allocate_trts}
\alias{allocate_trts}
\title{Define the possible allocation of treatments to units}
\usage{
allocate_trts(.edibble, ...)
}
\arguments{
\item{...}{One-sided or two-sided formula. If the input is a one-sided formula
then the whole treatment is applied to the specified unit.}

\item{.data}{An \code{edbl_graph} object.}

\item{class}{A sub-class. This is meant so that it can invoke the method
\code{randomise_trts.class}.}
}
\description{
This function adds the edges between level nodes in
an edibble graph that outlines the possible ways that
the treatment may be allocated to the recipient units.
}
\examples{
start_design() \%>\%
  set_units(block = 10,
            plot = nested_in(block, 3)) \%>\%
  set_trts(treat = c("A", "B", "C"),
           pest = c("a", "b")) \%>\%
  allocate_trts(treat ~ plot,
                pest ~ block)

# allocation works on edibble table too
lady_tasting_tea \%>\%
  edibble() \%>\%
  set_units(cup) \%>\%
  set_trts(first) \%>\%
  allocate_trts(first ~ cup)

}
\seealso{
Other user-facing functions: 
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
