% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allot.R
\name{allot_table}
\alias{allot_table}
\title{A shorthand for allot, assign and serve}
\usage{
allot_table(
  .design,
  ...,
  .order = "random",
  .seed = NULL,
  .constrain = nesting(.design)
)
}
\description{
A shorthand for allot, assign and serve
}
