% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomy.R
\name{anatomy}
\alias{anatomy}
\title{Anatomy of the design}
\usage{
anatomy(.edibble, ...)
}
\arguments{
\item{.edibble}{A complete edibble design object or edibble table.}

\item{...}{Any other arguments parsed to \code{dae::designAnatomy}.}
}
\value{
An object of class "des_anatomy".
}
\description{
This is a convenient wrapper for \code{dae::designAnatomy} where the \code{formulae} structure is
automatically determined by the unit and treatment structure specified in \code{edibble} system.
Note: the computation may be long if the design is quite complicated or there are many units.
}
