% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert an edibble data frame to normal data frame}
\usage{
as_data_frame(.data)
}
\arguments{
\item{.data}{can be a list or data frame}
}
\value{
A data.frame.
}
\description{
A patch function where there is an issue with edbl factors
}
