% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_tibble.edbl_table}
\alias{as_tibble.edbl_table}
\title{Convert an edibble data frame to normal data frame}
\usage{
\method{as_tibble}{edbl_table}(x, ...)
}
\arguments{
\item{x}{can be a list or data frame}

\item{...}{Not currently used.}
}
\value{
A data.frame.
}
\description{
A patch function where there is an issue with edbl factors
}
