% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{as_value}
\alias{as_value}
\title{Validation values}
\usage{
as_value(
  operator = c("=", "==", ">=", "<=", "<", ">", "!="),
  value,
  between,
  not_between
)
}
\arguments{
\item{value}{An optional value related to operator}

\item{between, not_between}{An optional numerical vector of size two where the
first entry is the minimum value and the second entry is the maximum value.
For \code{between}, the value is valid if within the range of minimum and maximum
value inclusive. For \code{not_between}, the value must lie outside of these values.
@return A list with two elements \code{operator} and \code{value}.}
}
\description{
Validation values
}
