% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{assign_trts}
\alias{assign_trts}
\title{Assign treatments to units}
\usage{
assign_trts(
  .design,
  order = "random",
  seed = NULL,
  constrain = nesting(.design)
)
}
\arguments{
\item{.design}{An edibble design which should have units, treatments and allotment defined.}

\item{.order}{A character vector signifying the apportion of treatments to units.
The value should be either "random", "systematic" or "systematic-random".
"random" allocates the treatment randomly to units based on specified allotment with restrictions
implied by unit structure.
"systematic" allocates the treatment in a systematic order to units.
"systematic-random" allocates the treatment in a systematic order to units but
where it is not possible to divide treatments equally (as the number of units are not divisible
by the number of levels of the treatment factor), then the extras are chosen randomly.}

\item{.seed}{A scalar value used to set the seed so that the result is reproducible.}

\item{.constrain}{The nesting structure for units.}
}
\description{
This function assigns specific treatment levels to actual units.
}
