% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{autofill_rcrds}
\alias{autofill_rcrds}
\title{Autofill the records}
\usage{
autofill_rcrds(.data, ..., .seed = NULL, .nsim = 1L)
}
\arguments{
\item{.data}{An edibble data.}

\item{...}{If supplied, it is a name-value pair where the name should
correspond to the record factor name and value is the f}

\item{.seed}{The seed number.}

\item{.nsim}{The number of simulations to run.}
}
\description{
This function fills the values of the record factors by automatically
choosing a simulation process. It tries to be smart by ensuring to use
values that is within expectation.
}
