% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{calculate}
\alias{calculate}
\title{Specify the calculation to derive variables}
\usage{
calculate(.x, .f, ..., .group_by = NULL)
}
\arguments{
\item{.x, .y}{Name of other variables in data.}

\item{.f}{A function, formula, or excel formula.}

\item{...}{Arguments to be passed into \code{.f}.}

\item{.l}{A vector of variable names.}
}
\value{
A special derivation class.
}
\description{
This function specifies the excel formula that should
be stored when the design is exported with \code{\link[=export_design]{export_design()}}.
The functions must be translatable to excel formula. Mappings
for some complex functions may not work.
}
