% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\docType{data}
\name{classics}
\alias{classics}
\alias{nclassics}
\title{A list of classical experimental designs}
\format{
A list of some classical experimental designs as either \code{edbl_graph}
or \code{edbl_df} object. The designs included are:
\describe{
\item{CRD} Completely randomised design.
\item{RCBD} Randomised complete block design.
\item{SPD} Split plot design.
}

An object of class \code{list} of length 4.
}
\usage{
nclassics

classics
}
\description{
A list of classical experimental designs
}
\keyword{datasets}
