% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster units together}
\usage{
cluster(.unit, ..., .vname)
}
\arguments{
\item{.unit}{The name of the child unit to group.}

\item{...}{a single number OR a sequence of two-sided formula where the
left-hand side corresponds to the name of the level (or the level number) of \code{.unit}
and the right-hand side is an integer specifying the number of levels nested under the
corresponding levels.}

\item{.vname}{the parent variable name}
}
\value{
Returns an isolated graph
}
\description{
Cluster units together
}
