% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{column}
\alias{column}
\title{Select a column.}
\usage{
column(x)
}
\arguments{
\item{x}{The column to select. Can be unquoted name or the column index.}
}
\description{
This is a helper function to select a column when data is supplied
for \code{lvls}.
}
