% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context-manipulations}
\alias{context-manipulations}
\alias{sort_context}
\alias{suppress_context}
\title{Context manipulations}
\usage{
sort_context(.data, ...)

suppress_context(.data, ...)
}
\arguments{
\item{.data}{An edibble design object.}
}
\description{
Context manipulations
}
