% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{cook_design}
\alias{cook_design}
\title{Cook the design in the kitchen}
\usage{
cook_design(x)
}
\arguments{
\item{x}{An edibble object.}
}
\value{
A Kitchen object.
}
\description{
This is a developer function to create a new Kitchen class with
the existing design.
}
\examples{
cook_design(takeout())
}
