% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{create_classic}
\alias{create_classic}
\alias{create_classic_code}
\title{Create some classical named experimental designs}
\usage{
create_classic(
  name = c("crd", "rcbd", "split"),
  r = NULL,
  t = NULL,
  n = NULL,
  t1 = t,
  t2 = NULL,
  b = r,
  seed = 2020,
  output = TRUE,
  code = output,
  info = output,
  print = output,
  ...
)

create_classic_code(
  name = c("crd", "rcbd", "spd"),
  r = NULL,
  t = NULL,
  n = NULL,
  t1 = t,
  t2 = NULL,
  b = r,
  seed = 2020,
  ...,
  quiet = FALSE
)
}
\arguments{
\item{name}{The short name of the classical named experiment design. See
under Details for the available named designs.}

\item{r}{The total number of replicates if applicable.}

\item{t}{The number of treatments testing.}

\item{n}{The total number of observations if applicable.}

\item{t1}{The number of levels for the first treatment factor.}

\item{t2}{The number of levels for the second treatment factor.}

\item{b}{The total number of blocks if applicable.}

\item{output}{Show the info, code and print outputs? Default is TRUE.}

\item{code}{Show the edibble code for generating the design. Default is
TRUE.}

\item{info}{Show information on design. Default is TRUE.}

\item{print}{Show the print out of the return object. Default is TRUE.}
}
\description{
This function creates some classical named experimental designs.
}
\details{
For The available named designs are:
\itemize{
\item "crd": completely randomised design
\item "rcbd": randomised complete block design
\item "split": split plot design
}
}
\examples{
create_classic("crd", n = 50, t = 5)

}
