% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{create_classics}
\alias{create_classics}
\title{Create some classical named experimental designs}
\usage{
create_classics(
  name = c("crd", "rcbd", "spd"),
  r = NULL,
  t = NULL,
  n = NULL,
  t1 = t,
  t2 = NULL,
  code = TRUE
)
}
\arguments{
\item{name}{The name of the classic experiment design. See
under Details for the available named designs.}

\item{r}{The total number of replicates if applicable.}

\item{t}{The number of treatments testing.}

\item{n}{The total number of observations if applicable.}

\item{t1}{The number of levels for the first treatment factor.}

\item{t2}{The number of levels for the second treatment factor.}

\item{code}{Show the edibble code for generating the design. Default is
TRUE.}

\item{b}{The total number of blocks if applicable.}
}
\description{
This function creates some classical experimental designs.
}
\details{
For The available named designs are:
\itemize{
\item "crd": completely randomised design
\item "rcbd": randomised complete block design
}
}
