% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{cross_edge_seq}
\alias{cross_edge_seq}
\title{Make an edge sequence for every pair from two vectors}
\usage{
cross_edge_seq(.nexus, vnames_from, vnames_to)
}
\arguments{
\item{.nexus}{An edibble nexus.}

\item{vnames_from}{A character vector of vertex names.}

\item{vnames_to}{A character vector of vertex names.}
}
\value{
An edge sequence (vector of vertex indices) that connects every node
from first vector to every node in the second vector.V()
}
\description{
Given two vectors of node names, return the edge sequence that connects
every node from first vector to the second vector.
}
