% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{crossed_by}
\alias{crossed_by}
\title{Specify the units to cross to index a new unit}
\usage{
crossed_by(..., attrs = NULL)
}
\arguments{
\item{...}{a sequence of units}

\item{attrs}{Currently not implemented.}
}
\value{
An object of class "cross_lvls".
}
\description{
\code{crossed_by(A, B)} is the same as \code{~A:B} but \code{crossed_by} offers more control over the names of the new units
as well as adding new attributes.
}
\examples{
design("Strip-Plot Design | Strip-Unit Design") \%>\%
  set_units(block = 3,
            row = nested_in(block, 7),
            col = nested_in(block, 6),
            unit = nested_in(block, crossed_by(row, col)))
}
