% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{derive_vars}
\alias{derive_vars}
\title{Derive variables from other variables.}
\usage{
derive_vars(.edibble, ...)
}
\description{
This is used to specify the excel formula for variables that
are derived based on other variables.
}
\seealso{
See \code{\link[=calculate]{calculate()}}, \code{\link[=calculate2]{calculate2()}} and \code{\link[=pcalculate]{pcalculate()}} to
specify the excel formula.
}
