% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_model}
\alias{design_model}
\title{A baseline model for given experimental design}
\usage{
design_model(data, type = c("anova", "lmer"))
}
\arguments{
\item{data}{An edibble data.}

\item{type}{The type of model expression to return.}
}
\description{
This
}
