% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ed_levels}
\alias{ed_levels}
\title{Get all the levels of edibble variables as a list}
\usage{
ed_levels(x)
}
\arguments{
\item{x}{An edibble graph or edibble object.}
}
\value{
Returns a named list of levels of all edibble variables.
}
\description{
Get all the levels of edibble variables as a list
}
