% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble-package.R
\docType{package}
\name{edibble-package}
\alias{edibble-package}
\alias{_PACKAGE}
\title{edibble: Generate experimental designs using the grammar of experimental design}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A system to facilitate statistical thinking of 
    adapting experimental designs to different conditions.
}
\details{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\strong{(WIP)}
}
\section{Website}{

\itemize{
\item The website for the package is at \url{https://edibble.emitanaka.org}
}
\itemize{
\item Discussion is at \url{https://github.com/emitanaka/edibble/discussions}
}
}

\section{Package options}{

The following options are used for changing the default view for edibble
graph or edibble design plot:
\itemize{
\item \code{edibble.vertex.label.color.default}
\item \code{edibble.vertex.label.color.unit}
}

The following options are used for changing the default view for the print
out of edibble design or edibble graph.
\itemize{
\item \code{edibble.tree.decorate.trts}
\item \code{edibble.tree.decorate.units}
\item \code{edibble.tree.decorate.rcrd}
\item \code{edibble.tree.decorate.levels}
\item \code{edibble.tree.decorate.main}
}

TODO
}

\seealso{
Useful links:
\itemize{
  \item \url{https://edibble.emitanaka.org/}
  \item \url{https://github.com/emitanaka/edibble}
  \item Report bugs at \url{https://github.com/emitanaka/edibble/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
