% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{edibble}
\alias{edibble}
\title{Make input edibble}
\usage{
edibble(x, ..., units = NULL, trts = NULL)
}
\arguments{
\item{x}{A named list of edibble variables or data frame.}

\item{...}{Passed to \code{new_tibble}.}

\item{units}{A character vector.}
}
\value{
A data frame of class \code{edbl}.
}
\description{
If variables are already defined as external data then you can import this
data as edibble. T
}
\seealso{
See \code{\link[=initiate_design]{initiate_design()}} for initiating design from scratch.
}
