% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{EdibbleDesign}
\alias{EdibbleDesign}
\title{An R6 Class for an edibble design}
\description{
An object with EdibbleDesign holds the core information related to
the experimental design.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{the name}

\item{\code{graph}}{the name}

\item{\code{table}}{the name}

\item{\code{context}}{the name}

\item{\code{muffle}}{the nam}

\item{\code{active}}{Which is active}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EdibbleDesign$new()}}
\item \href{#method-finalize}{\code{EdibbleDesign$finalize()}}
\item \href{#method-plot}{\code{EdibbleDesign$plot()}}
\item \href{#method-print}{\code{EdibbleDesign$print()}}
\item \href{#method-save_seed}{\code{EdibbleDesign$save_seed()}}
\item \href{#method-clone}{\code{EdibbleDesign$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise the edibble design. See also \code{\link[=start_design]{start_design()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$new(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name or title of the project. The title
should be informative and not overly long.
If not supplied the default is "An edibble design".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Invoked when the EdibbleDesign object is deleted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the intermediate contruct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$plot(view = c("high", "low"), ..., main = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view}}{A high- or -low level view.}

\item{\code{...}}{Arguments passed into \code{plot.igraph}.}

\item{\code{main}}{The title of the plot. By default it is the
name of the EdibbleDesign object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the intermediate construct.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed into \code{print}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_seed"></a>}}
\if{latex}{\out{\hypertarget{method-save_seed}{}}}
\subsection{Method \code{save_seed()}}{
This saves the \code{.Random.seed} used just prior to the
randomisation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$save_seed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdibbleDesign$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
