% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{eval_traits}
\alias{eval_traits}
\title{Evaluate traits}
\usage{
eval_traits(x, vname, graph = NULL)
}
\arguments{
\item{x}{A value that may be a single number, unnamed vector,
a one sided formula or linkabble.}

\item{vname}{The name of the variable.}

\item{graph}{An edibble graph with levels converted as list used
for evaluation in any expressions.}
}
\value{
Returns an evaluated expression.
}
\description{
Evaluate traits
}
