% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{examine_recipe}
\alias{examine_recipe}
\title{Check the recipe code of a takeout}
\usage{
examine_recipe(takeout)
}
\arguments{
\item{takeout}{A \code{takeout} object.}
}
\value{
A named design object.
}
\description{
Check the recipe code of a takeout
}
