% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{exp_unit_names}
\alias{exp_unit_names}
\title{Get hte experimental unit names}
\usage{
exp_unit_names(.edibble)
}
\description{
If there are more than one treatment factor applied to multiple units, then the
return object will be a character vector.
}
