% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{expect_rcrds}
\alias{expect_rcrds}
\title{Set the expected values for recording variables}
\usage{
expect_rcrds(.edibble, ...)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_table}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Name-value pairs with the name belonging to the variable
that are plan to be recorded from \code{set_rcrds()} and the values are
the expected types and values set by helper functions, see \code{?expect-rcrds}.}
}
\description{
Set the expected values for recording variables
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
