% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{expect_vars}
\alias{expect_vars}
\title{Set the expected values for variables}
\usage{
expect_vars(.design, ...)
}
\arguments{
\item{...}{Name-value pairs with the name belonging to the variable
that are plan to be recorded from \code{record_vars()} and the values are
the expected types and values set by helper functions, see \code{?expect-vars}.}
}
\description{
Set the expected values for variables
}
