% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{extract-lvl-nodes}
\alias{extract-lvl-nodes}
\alias{$.edbl_lnodes}
\alias{[.edbl_lnodes}
\alias{[<-.edbl_lnodes}
\alias{[[.edbl_lnodes}
\alias{[[<-.edbl_lnodes}
\title{Extract or replace parts of the level nodes}
\usage{
\method{$}{edbl_lnodes}(x, name)

\method{[}{edbl_lnodes}(x, i, ...)

\method{[}{edbl_lnodes}(x, i, ...) <- value

\method{[[}{edbl_lnodes}(x, i, ...)

\method{[[}{edbl_lnodes}(x, i, ...) <- value
}
\arguments{
\item{x}{An edibble level nodes.}

\item{name}{The name to extract.}

\item{i}{The index.}

\item{value}{The value to replace with.}
}
\description{
The level nodes are stored as a named list of nodes where the name
corresponds to the id of the corresponding factor. This makes the
access of level nodes slightly awkward. For example, to extract the
id of the level nodes, you have to iterate over every list.
}
\examples{
crd <- takeout(menu_crd())

}
