% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{fct}
\alias{fct}
\alias{fct_attrs}
\title{Setting the traits of factors}
\usage{
fct(.levels, ...)

fct_attrs(.levels, ...)
}
\arguments{
\item{.levels}{Either a short hand given as either as a single integer (number of levels),
a vector or levels created from \code{lvls()}.}

\item{...}{A name-value pair of attributes. The value must be a scalar and
attributed to the whole factor (not individual levels).
The values are added as attributes to the output object.}
}
\description{
This function is used to set characteristics of the factors.
}
\examples{
fct(c("A", "B"))
}
\seealso{
lvls
}
