% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{fct_attrs}
\alias{fct_attrs}
\title{Setting the traits of factors}
\usage{
fct_attrs(
  levels = NULL,
  label = NULL,
  description = NULL,
  unit_of_measure = NULL,
  class = NULL,
  ...
)
}
\arguments{
\item{levels}{An \code{edbl_levels} object that should contain information about the levels
in the factor.}

\item{label}{A string that denotes the long name of the factor.}

\item{description}{The text description of the factor.}

\item{unit_of_measure}{A string denoting the unit of measurement if applicable.}

\item{class}{An optional subclass.}

\item{...}{A name-value pair of attributes. The value must be a scalar and
attributed to the whole factor (not individual levels).
The values are added as attributes to the output object.}
}
\value{
An \code{edbl_levels} object.
}
\description{
This function is used to set characteristics of the factors.
}
\seealso{
lvl_traits
}
