% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{fct_generator}
\alias{fct_generator}
\title{Factor name generator}
\usage{
fct_generator(labels, nlevels)
}
\arguments{
\item{labels}{A character vector specifying the custom labels for the factor levels.}

\item{nlevels}{An integer or a vector of integers indicating the number of repetitions for each label.
If a single integer is provided, it is recycled to match the length of \code{labels}.
If a vector is provided, it should have the same length as \code{labels}.}
}
\value{
A factor with custom levels and repetitions.
}
\description{
Generate a factor with custom levels and repetitions.
}
\details{
This function creates a factor with custom labels and specified repetitions for each label.
}
\examples{
# Example usage of the function
fct_generator(labels = c("A", "B", "C"), nlevels = 3)

}
