% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{find_classical_names}
\alias{find_classical_names}
\title{Find the short names of the classical named designs}
\usage{
find_classical_names(pkgs = NULL)
}
\arguments{
\item{pkgs}{A character vector containing the package names to search classical
named designs from. By default it will search edibble and other packages loaded.}
}
\description{
Find the short names of the classical named designs
}
