% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-input.R
\name{graph_input}
\alias{graph_input}
\title{A function to process input as input for graph manipulation}
\usage{
graph_input(input, prov, ...)
}
\arguments{
\item{input}{An input.}

\item{prov}{A provenance object.}

\item{...}{Unused.}
}
\description{
A function to process input as input for graph manipulation
}
