% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{initiate_design}
\alias{initiate_design}
\title{Initiate the edibble design}
\usage{
initiate_design(name = NULL)
}
\arguments{
\item{name}{Optional name used as title for printing the design.}
}
\value{
An \code{edbl_nexus} object which is a special type of \code{igraph} class.
}
\description{
This function doesn't really do much besides create an empty igraph.
}
\examples{
initiate_design("My design")
}
\seealso{
Add variables to this design with \code{\link[=set_units]{set_units()}}, \code{\link[=set_trts]{set_trts()}}, and
\code{\link[=measure_units]{measure_units()}}.
}
