% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{iplot}
\alias{iplot}
\alias{iplot.EdibbleDesign}
\alias{iplot.edbl_graph}
\alias{iplot.edbl_table}
\title{Interactive plot of the graph of an edibble design}
\usage{
iplot(.edibble, ...)

\method{iplot}{EdibbleDesign}(.edibble, view = c("low", "high"), ..., width = 450, height = 450)

\method{iplot}{edbl_graph}(.edibble, view = c("low", "high"), ..., width = 450, height = 450)

\method{iplot}{edbl_table}(.edibble, view = c("low", "high"), ..., width = 450, height = 450)
}
\arguments{
\item{.edibble}{An edibble desgin, an edibble table or an edibble graph.}

\item{...}{Additional parameters that are passed into \code{igraph::tkplot}.}

\item{width}{The width of the canvas. Same as the \code{canvas.width} in \code{tkplot}.}

\item{height}{The height of the canvas. Same as the \code{canvas.height} in \code{tkplot}.}
}
\description{
This is similar to \code{plot.EdibbleDesign()} except it is an interactive version
using \code{igraph::tkplot()} behind the scenes. Unlike \code{plot.EdibbleDesign()},
the default view is "low".
}
\seealso{
See \code{\link[=plot.edbl_graph]{plot.edbl_graph()}} for static plot of the graph of an edibble design.
}
