% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{is_edibble}
\alias{is_edibble}
\title{Test if the object is edibble}
\usage{
is_edibble(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits the \code{edbl} class.
}
\description{
A simple function that return \code{TRUE} if the object inherits
the \code{edbl} class, \code{FALSE} otherwise.
}
\seealso{
See \code{\link[=is_edibble_nexus]{is_edibble_nexus()}} for testing \code{edbl_nexus}.
}
