% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{is_edibble_graph}
\alias{is_edibble_graph}
\title{Test if the object is an edibble graph}
\usage{
is_edibble_graph(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits the \code{edbl_graph} class.
}
\description{
A simple function that return \code{TRUE} if the object inherits
the \code{edbl_graph} class, \code{FALSE} otherwise.
}
\seealso{
See \code{\link[=is_edibble]{is_edibble()}} for testing \code{edbl_df} data frame.
}
