% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{is_provenance}
\alias{is_provenance}
\title{Check if an object is an instance of the "Provenance" class.}
\usage{
is_provenance(x)
}
\arguments{
\item{x}{An object to be checked for its class membership.}
}
\value{
\code{TRUE} if the object is an instance of the "Provenance" class,
\code{FALSE} otherwise.
}
\description{
This function determines whether the given object is an instance of the
"Provenance" class.
}
