% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{is_takeout}
\alias{is_takeout}
\title{A function to check if the output is a takeout design}
\usage{
is_takeout(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value.
}
\description{
The function returns \code{TRUE} if the input is a takeout design.
}
\examples{
is_takeout(takeout())
}
