% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_nested}
\alias{label_nested}
\alias{label_distinct}
\alias{index_levels}
\title{Label with nested or distinct labels}
\usage{
label_nested(x)

label_distinct(x)

index_levels(x)
}
\arguments{
\item{x}{A unit vector.}
}
\description{
Label with nested or distinct labels
}
