% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkabble.R
\name{linkabble}
\alias{linkabble}
\title{Make a link table}
\usage{
linkabble(flist, .allocate = 1, .type = "nested")
}
\arguments{
\item{.allocate}{the number of units to allocate for every level.
It may be a vector but the length needs to match the order and
the number of combined levels of given trait variables.}

\item{...}{trait variables}
}
\description{
Make a link table
}
\seealso{
set_link
}
