% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{measure_responses}
\alias{measure_responses}
\title{Measure response of given units}
\usage{
measure_responses(
  .design,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{...}{Name-value pair.}

\item{.data}{An \code{edbl_graph} object}
}
\description{
This function creates new nodes to edibble graph with the name
corresponding to the intended response that will be measured.
}
