% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_crd}
\alias{menu_crd}
\title{Completely randomised design}
\usage{
menu_crd(
  t = random_integer_small(),
  n = random_integer_medium(min = t),
  r = NULL,
  seed = random_seed_number(),
  name_title = "Completely Randomised Design",
  name_unit = "unit",
  name_trt = "trt"
)
}
\arguments{
\item{t}{The number of treatment levels}

\item{n}{The number of experimental units}

\item{r}{(Optional) The number of replicates.}

\item{seed}{A scalar value for computational reproducibility.}

\item{name_title}{The name of the design or title.}

\item{name_unit}{The name of the unit factor}

\item{name_trt}{The name of the treatment factor.}
}
\value{
A recipe for completely randomised design.
}
\description{
Completely randomised design
}
\examples{
menu_crd(t = 3, n = 10)
}
\seealso{
Other recipe-designs: 
\code{\link{menu_bibd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_hyper_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_split}()},
\code{\link{menu_strip}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
