% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{menu_lsd}
\alias{menu_lsd}
\title{Prepare classical Latin square design}
\usage{
menu_lsd(t = random_integer_small(), seed = random_seed_number())
}
\arguments{
\item{t}{The number of treatments}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare classical Latin square design
}
\seealso{
Other named-designs: 
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_split}()},
\code{\link{menu_youden}()}
}
\concept{named-designs}
