% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_rcbd}
\alias{menu_rcbd}
\title{Prepare a randomised complete block design}
\usage{
menu_rcbd(
  t = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number(),
  name_title = "Randomised Complete Block Design",
  name_block = "block",
  name_unit = "unit",
  name_trt = "trt"
)
}
\arguments{
\item{t}{The number of treatments.}

\item{r}{The number of replications for each treatment level.}

\item{seed}{A scalar value for computational reproducibility.}

\item{name_title}{The name of the design or title.}

\item{name_block}{The name of the block factor.}

\item{name_unit}{The name of the unit factor}

\item{name_trt}{The name of the treatment factor.}
}
\value{
A recipe for randomised complete block design.
}
\description{
Prepare a randomised complete block design
}
\examples{
menu_rcbd(t = 3, r = 2)
}
\seealso{
Other recipe-designs: 
\code{\link{menu_bibd}()},
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_hyper_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_split}()},
\code{\link{menu_strip}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
