% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_split}
\alias{menu_split}
\title{Split-unit design}
\usage{
menu_split(
  t1 = random_integer_small(),
  t2 = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number(),
  name_title = c("Split-Plot Design", "Split-Unit Design"),
  name_mainplot = "mainplot",
  name_subplot = "subplot",
  name_trt1 = "trt1",
  name_trt2 = "trt2"
)
}
\arguments{
\item{t1}{The number of treatment levels for the main plots.}

\item{t2}{The number of treatment levels for the subplots.}

\item{r}{The number of replications for each treatment level.}

\item{seed}{A scalar value for computational reproducibility.}

\item{name_title}{The name of the design or title.}
}
\value{
A recipe split-plot design.
}
\description{
Originally referred to as split-plot design when it was first used.
}
\examples{
menu_split(t1 = 3, t2 = 2, r = 4)
}
\seealso{
Other recipe-designs: 
\code{\link{menu_bibd}()},
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_hyper_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_strip}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
