% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_strip}
\alias{menu_strip}
\title{Strip-unit design}
\usage{
menu_strip(
  t1 = random_integer_small(),
  t2 = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number(),
  name_title = c("Strip-Plot Design", "Strip-Unit Design"),
  name_block = "block",
  name_unit = "unit",
  name_row = "row",
  name_col = "col",
  name_trt1 = "trt1",
  name_trt2 = "trt2"
)
}
\arguments{
\item{t1}{The number of treatment levels for the main plots.}

\item{t2}{The number of treatment levels for the subplots.}

\item{r}{The number of replications for each treatment level.}

\item{seed}{A scalar value for computational reproducibility.}

\item{name_title}{The name of the design or title.}

\item{name_unit}{The name of the unit factor}

\item{name_row, name_col}{The names for the row and column factors.}

\item{name_trt1, name_trt2}{The name of the treatment factors.}
}
\value{
A recipe strip-unit design.
}
\description{
Strip-unit design
}
\examples{
menu_strip(t1 = 3, t2 = 3, r = 2)
}
\seealso{
Other recipe-designs: 
\code{\link{menu_bibd}()},
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_hyper_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_split}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
