% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{n_units}
\alias{n_units}
\alias{n_units.edbl_nexus}
\alias{n_units.edbl_df}
\title{Number of units associated with the given variable}
\usage{
n_units(.data, ...)

\method{n_units}{edbl_nexus}(.nexus, var = NULL)

\method{n_units}{edbl_df}(.data, var)
}
\arguments{
\item{.nexus, .data}{An edibble nexus or an edibble data frame.}

\item{var}{The name of the edibble variable. If this is not supplied then
the number of experimental unit is returned.}
}
\description{
Number of units associated with the given variable
}
