% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{NamedDesign}
\alias{NamedDesign}
\title{An R6 Class for a named experimental design}
\description{
An R6 Class for a named experimental design

An R6 Class for a named experimental design
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The short name of the experimental design.}

\item{\code{name_full}}{The full name of the experimental design.}

\item{\code{code}}{The code, based on the grammar of experimental design,
to generate the experimental design.}

\item{\code{n}}{The total number of experimental units.}

\item{\code{t}}{The total number of treatments tested.}

\item{\code{info_always}}{A character vector containing information or properties
that is always true for this named design.  Each entry in the vector forms
a dot point.}

\item{\code{info_this}}{A character vector containing information or properties
that is true for this named design for the chosen parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NamedDesign$new()}}
\item \href{#method-add_code}{\code{NamedDesign$add_code()}}
\item \href{#method-add_seed}{\code{NamedDesign$add_seed()}}
\item \href{#method-add_info}{\code{NamedDesign$add_info()}}
\item \href{#method-final_code}{\code{NamedDesign$final_code()}}
\item \href{#method-decorate_units}{\code{NamedDesign$decorate_units()}}
\item \href{#method-decorate_trts}{\code{NamedDesign$decorate_trts()}}
\item \href{#method-clone}{\code{NamedDesign$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise the named design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$new(name, name_full = name, n, t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Short name.}

\item{\code{name_full}}{The full name of the design.}

\item{\code{n}}{Total number of experimental units.}

\item{\code{t}}{Total number of treatments tested.}

\item{\code{seed}}{The seed number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_code"></a>}}
\if{latex}{\out{\hypertarget{method-add_code}{}}}
\subsection{Method \code{add_code()}}{
Append code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$add_code(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A string specifying the code to append.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_seed"></a>}}
\if{latex}{\out{\hypertarget{method-add_seed}{}}}
\subsection{Method \code{add_seed()}}{
Add code to set seed at the beginning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$add_seed(seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The seed number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_info"></a>}}
\if{latex}{\out{\hypertarget{method-add_info}{}}}
\subsection{Method \code{add_info()}}{
Add information about the named design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$add_info(x, always = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A string specifying one info about a design.}

\item{\code{always}}{A logical to specify whether the added information is
always true for this this design or only for the given parameter.
Default is FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-final_code"></a>}}
\if{latex}{\out{\hypertarget{method-final_code}{}}}
\subsection{Method \code{final_code()}}{
This adds the final bit of code needed to generate the edibble
data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$final_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decorate_units"></a>}}
\if{latex}{\out{\hypertarget{method-decorate_units}{}}}
\subsection{Method \code{decorate_units()}}{
A function to decorate units.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$decorate_units(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A string to decorate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decorate_trts"></a>}}
\if{latex}{\out{\hypertarget{method-decorate_trts}{}}}
\subsection{Method \code{decorate_trts()}}{
A function to decorate treatments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$decorate_trts(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A string to decorate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedDesign$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
