% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names-ed-vars}
\alias{names-ed-vars}
\alias{names_units}
\alias{names_trts}
\alias{names_resp}
\alias{names_vars}
\alias{names_vars.edbl_graph}
\alias{names_vars.edbl_df}
\alias{names.edbl_graph}
\alias{names<-.edbl_graph}
\title{Names of variables}
\usage{
names_units(.object)

names_trts(.object)

names_resp(.object)

names_vars(.object, ...)

\method{names_vars}{edbl_graph}(.data, class = NULL)

\method{names_vars}{edbl_df}(.data, class = NULL)

\method{names}{edbl_graph}(.data, label = TRUE)

\method{names}{edbl_graph}(x) <- value
}
\arguments{
\item{.object, .data, .data}{Either an edibble data frame or edibble graph.}

\item{class}{A string.}
}
\value{
A character vector of the names of the variables with the given class.
}
\description{
Functions to get or set the names of variables (with particular class) from
edibble or edibble graph object.
}
\examples{
names_units(nclassics$split)
names_trts(nclassics$split)
}
