% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names_by_class}
\alias{names_by_class}
\title{Get names by vertex class}
\usage{
names_by_class(.graph, class = NULL)
}
\arguments{
\item{class}{The class name.}
}
\description{
Get the name of variables by class.
}
