% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nesting_structure}
\alias{nesting_structure}
\title{Get the nesting structure for the units}
\usage{
nesting_structure(design)
}
\arguments{
\item{design}{An edibble design}
}
\value{
Return a named list. Only shows the direct parent.
}
\description{
Get the nesting structure for the units
}
\examples{
nesting_structure(takeout(menu_split()))
}
