% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{new_edibble}
\alias{new_edibble}
\alias{as_edibble}
\alias{edibble}
\title{An edibble table constructor}
\usage{
new_edibble(.data, ..., design = NULL, class = NULL)

as_edibble(.data, ...)

edibble(
  .data,
  title = NULL,
  name = "edibble",
  .record = TRUE,
  seed = NULL,
  provenance = Provenance$new(),
  ...
)
}
\arguments{
\item{.data}{data frame or list of the same size.}

\item{...}{Passed to \code{new_tibble}.}

\item{design}{An edibble graph object.}

\item{class}{Subclasses for edibble table. The default is NULL.}

\item{title}{Optional title of the experiment.}

\item{name}{Optional name of the experiment.}

\item{.record}{A logical value. This indicates whether to record this
code step. The default is TRUE. It should remain TRUE unless this
function is used as a wrapper in other code.}

\item{seed}{A seed number for reproducibility.}

\item{provenance}{An environment setup in a manner to store methods and
information to trace the origin of the design}
}
\value{
An edibble table.
}
\description{
This helps to construct a new edibble table which is a special type
of tibble.
}
