% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{new_edibble_fct}
\alias{new_edibble_fct}
\title{Constructor for an edibble variable}
\usage{
new_edibble_fct(
  labels = character(),
  levels = unique(labels),
  name = character(),
  rep = NULL,
  ...,
  class = NULL
)
}
\description{
Constructor for an edibble variable
}
