% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{new_edibble_var}
\alias{new_edibble_var}
\title{Constructor for an edibble variable}
\usage{
new_edibble_var(
  labels = character(),
  levels = unique(labels),
  name = character(),
  rep = NULL,
  ...,
  class = NULL
)
}
\description{
Constructor for an edibble variable
}
