% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_si_prefix}
\alias{number_si_prefix}
\title{Numbers with SI prefix}
\source{
https://en.wikipedia.org/wiki/Metric_prefix
}
\usage{
number_si_prefix(x)
}
\arguments{
\item{x}{A numeric vector to format with SI prefix}
}
\description{
It's called SI prefix but the letter notation is added as a suffix.
The largest prefix is yotta $10^{24}$.
}
