% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{order_trts}
\alias{order_trts}
\title{A custom ordering algorithm}
\usage{
order_trts(x, ...)
}
\arguments{
\item{x}{A string specifying the class}

\item{...}{Other arguments.}
}
\description{
A custom ordering algorithm
}
