% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{path_seq}
\alias{path_seq}
\title{Get the edge sequence that draws the apth.}
\usage{
path_seq(.data, vnames)
}
\arguments{
\item{.data}{An edibble graph.}

\item{vnames}{The vertex names.}
}
\description{
Get the edge sequence that draws the apth.
}
