% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_trts_widelist}
\alias{pivot_trts_widelist}
\title{Pivot treatments to a wider list or table format}
\usage{
pivot_trts_widelist(.data, trts = NULL)
}
\arguments{
\item{.data}{An edibble table.}

\item{trts}{A vector of character with the treatment names. By default it is
NULL and includes all the treatments.}
}
\description{
Pivot treatments to a wider list or table format
}
