% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edbl_design}
\alias{plot.edbl_design}
\alias{plot.edbl_table}
\title{Interactive plot of the edibble design}
\usage{
\method{plot}{edbl_design}(
  .design,
  view = c("factors", "levels"),
  width = "100\%",
  height = NULL,
  seed = 1,
  show_buttons = TRUE,
  ...
)

\method{plot}{edbl_table}(.table, ...)
}
\arguments{
\item{.design}{An edibble design.}

\item{view}{A string of either "factors" or "levels".}

\item{width, height}{The width and height of the plot.}

\item{seed}{A seed number so same plot is always generated.}

\item{show_buttons}{A logical value indicating whether to hide or show the buttons.}
}
\description{
Interactive plot of the edibble design
}
