% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edbl_nexus}
\alias{plot.edbl_nexus}
\title{Plot graph of edibble variables}
\usage{
\method{plot}{edbl_nexus}(
  .nexus,
  view = c("high", "low"),
  ...,
  layout = igraph::layout.auto,
  main = NULL
)
}
\arguments{
\item{.nexus}{An edibble nexus.}

\item{view}{A string specifying the level of the view to plot.}

\item{...}{Additional parameters passed to \code{plot.igraph}.}

\item{layout}{An igraph layout.}
}
\description{
Plot graph of edibble variables
}
\examples{
plot(.nexus, sample_levels(n = 30))

}
