% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edibble}
\alias{plot.edibble}
\alias{plot.EdibbleDesign}
\alias{plot.edbl_graph}
\alias{plot.edbl_table}
\title{Plot graph of an edibble design}
\usage{
\method{plot}{EdibbleDesign}(.edibble, view = c("high", "low"), ..., main = NULL)

\method{plot}{edbl_graph}(.edibble, view = c("high", "low"), ..., main = NULL)

\method{plot}{edbl_table}(.edibble, view = c("high", "low"), ..., main = NULL)
}
\arguments{
\item{.edibble}{An edibble desgin, an edibble table or an edibble graph.}

\item{view}{A string specifying the level of the view ("high" or "low") to plot.
By default it is "high".}

\item{...}{Additional parameters passed to \code{plot.igraph}.}

\item{main}{The title of the plot. By default it is the
name of the edibble design.}
}
\description{
There are two available view for the graph of edibble variables:
\itemize{
\item \code{"high"} level view where only nodes that are variables are shown, or
\item \code{"low"} level view where only nodes that are levels are shown.
The default view is "high" since the "low" view is usually too cluttered to
be of any use.
}

The main use for this is to visualise the relationship between variables
and levels interactively. The plot output is created using \code{plot.igraph()}.
As typical of network visualisation, the layout may require adjusting for a
better look.

This plot is not optimal for a polished publication. It is designed more for
a quick exploratory use.
}
\seealso{
See \code{\link[=iplot]{iplot()}} for interactive plot of the graph of an edibble design.
}
