% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.EdibbleDesign}
\alias{plot.EdibbleDesign}
\title{Plot graph of edibble variables}
\usage{
\method{plot}{EdibbleDesign}(.data, view = "high", ..., main = NULL)
}
\arguments{
\item{.data}{An edibble desgin.}

\item{view}{A string specifying the level of the view to plot.}

\item{...}{Additional parameters passed to \code{plot.igraph}.}

\item{main}{The title of the plot. By default it is\code{.data$.name}.}
}
\description{
Plot graph of edibble variables
}
