% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_factorial}
\alias{prep_classical_factorial}
\title{Prepare a factorial design}
\usage{
prep_classical_factorial(
  trt = c(sample(1:10, 1), sample(1:10, 1)),
  r = sample(1:10, 1),
  design = c("crd", "rcbd"),
  seed = sample(1000, 1)
)
}
\arguments{
\item{trt}{A vector of the number of levels for each treatment factor.}

\item{r}{The number of replicate blocks.}

\item{design}{The unit structure: "crd" or "rcbd". The default is "crd".}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare a factorial design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_rcbd}()},
\code{\link{prep_classical_split}()}
}
\concept{named-designs}
