% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_graeco}
\alias{prep_classical_graeco}
\title{Graeco-Latin Square Design}
\usage{
prep_classical_graeco(t = random_integer_small(), seed = random_seed_number())
}
\arguments{
\item{t}{The number of treatments.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Graeco-Latin Square Design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_lsd}()},
\code{\link{prep_classical_rcbd}()},
\code{\link{prep_classical_split}()},
\code{\link{prep_classical_youden}()}
}
\concept{named-designs}
