% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_lsd}
\alias{prep_classical_lsd}
\title{Prepare classical Latin square design}
\usage{
prep_classical_lsd(t = random_integer_small(), seed = random_seed_number())
}
\arguments{
\item{t}{The number of treatments}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare classical Latin square design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_graeco}()},
\code{\link{prep_classical_rcbd}()},
\code{\link{prep_classical_split}()},
\code{\link{prep_classical_youden}()}
}
\concept{named-designs}
