% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_rcbd}
\alias{prep_classical_rcbd}
\title{Prepare a randomised complete block design}
\usage{
prep_classical_rcbd(
  t = 1 + sample(10, 1),
  r = 1 + sample(10, 1),
  seed = sample(1000, 1)
)
}
\arguments{
\item{t}{The number of treatments.}

\item{r}{The number of replicate blocks.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare a randomised complete block design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_split}()}
}
\concept{named-designs}
