% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_youden}
\alias{prep_classical_youden}
\title{Youden square design}
\usage{
prep_classical_youden(
  nc = random_integer_small(),
  t = random_integer_small(min = nc + 1),
  seed = random_seed_number()
)
}
\arguments{
\item{t}{The number of treatments.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Youden square design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_graeco}()},
\code{\link{prep_classical_lsd}()},
\code{\link{prep_classical_rcbd}()},
\code{\link{prep_classical_split}()}
}
\concept{named-designs}
