% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{Provenance}
\alias{Provenance}
\title{A manipulator for the edbl_design.}
\description{
A manipulator for the edbl_design.

A manipulator for the edbl_design.
}
\details{
Internal functions should create a new Provenance object.
The Provenance contains a set of operations to manipulate the nodes and edges of
the edibble design object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fct_new_id}}{Get a new factor id.}

\item{\code{lvl_new_id}}{Get a new level id.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fct_nodes}}{Get the factor nodes}

\item{\code{lvl_nodes}}{Get the level nodes}

\item{\code{fct_edges}}{Get the factor edges}

\item{\code{lvl_edges}}{Get the level edges}

\item{\code{fct_n}}{Get the number of nodes in factor graph}

\item{\code{lvl_n}}{Get the number of nodes in level graph}

\item{\code{fct_leaves}}{Get the leave factor ids.}

\item{\code{rcrd_ids}}{Get the ids for all edbl_rcrd factors.}

\item{\code{unit_ids}}{Get the ids for all edbl_unit factors.}

\item{\code{trt_ids}}{Get the ids for all edbl_trt factors.}

\item{\code{trt_names}}{Get the node labels for treatments}

\item{\code{unit_names}}{Get the node labels for units}

\item{\code{rcrd_names}}{Get the node labels for record}

\item{\code{is_connected}}{Check if nodes are connected.}

\item{\code{fct_new_id}}{Get a new factor id.}

\item{\code{lvl_new_id}}{Get a new level id.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Provenance-new}{\code{Provenance$new()}}
\item \href{#method-Provenance-set_name}{\code{Provenance$set_name()}}
\item \href{#method-Provenance-reactivate}{\code{Provenance$reactivate()}}
\item \href{#method-Provenance-deactivate}{\code{Provenance$deactivate()}}
\item \href{#method-Provenance-fct_id_by_name}{\code{Provenance$fct_id_by_name()}}
\item \href{#method-Provenance-fct_id_by_class}{\code{Provenance$fct_id_by_class()}}
\item \href{#method-Provenance-lvl_id_by_value}{\code{Provenance$lvl_id_by_value()}}
\item \href{#method-Provenance-fct_names}{\code{Provenance$fct_names()}}
\item \href{#method-Provenance-lvl_names}{\code{Provenance$lvl_names()}}
\item \href{#method-Provenance-append_fct_nodes}{\code{Provenance$append_fct_nodes()}}
\item \href{#method-Provenance-append_lvl_nodes}{\code{Provenance$append_lvl_nodes()}}
\item \href{#method-Provenance-append_fct_edges}{\code{Provenance$append_fct_edges()}}
\item \href{#method-Provenance-append_lvl_edges}{\code{Provenance$append_lvl_edges()}}
\item \href{#method-Provenance-fct_class}{\code{Provenance$fct_class()}}
\item \href{#method-Provenance-lvl_class}{\code{Provenance$lvl_class()}}
\item \href{#method-Provenance-fct_child}{\code{Provenance$fct_child()}}
\item \href{#method-Provenance-lvl_child}{\code{Provenance$lvl_child()}}
\item \href{#method-Provenance-lvl_id_by_class}{\code{Provenance$lvl_id_by_class()}}
\item \href{#method-Provenance-fct_parent}{\code{Provenance$fct_parent()}}
\item \href{#method-Provenance-lvl_parent}{\code{Provenance$lvl_parent()}}
\item \href{#method-Provenance-fct_ancestor}{\code{Provenance$fct_ancestor()}}
\item \href{#method-Provenance-lvl_ancestor}{\code{Provenance$lvl_ancestor()}}
\item \href{#method-Provenance-fct_levels}{\code{Provenance$fct_levels()}}
\item \href{#method-Provenance-fct_exists}{\code{Provenance$fct_exists()}}
\item \href{#method-Provenance-trts_exists}{\code{Provenance$trts_exists()}}
\item \href{#method-Provenance-units_exists}{\code{Provenance$units_exists()}}
\item \href{#method-Provenance-rcrds_exists}{\code{Provenance$rcrds_exists()}}
\item \href{#method-Provenance-save_seed}{\code{Provenance$save_seed()}}
\item \href{#method-Provenance-get_history}{\code{Provenance$get_history()}}
\item \href{#method-Provenance-get_seed}{\code{Provenance$get_seed()}}
\item \href{#method-Provenance-get_session_info}{\code{Provenance$get_session_info()}}
\item \href{#method-Provenance-get_edibble_version}{\code{Provenance$get_edibble_version()}}
\item \href{#method-Provenance-record_step}{\code{Provenance$record_step()}}
\item \href{#method-Provenance-record_history_external}{\code{Provenance$record_history_external()}}
\item \href{#method-Provenance-record_history_internal}{\code{Provenance$record_history_internal()}}
\item \href{#method-Provenance-add_history_internal}{\code{Provenance$add_history_internal()}}
\item \href{#method-Provenance-clone}{\code{Provenance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$new(graph = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design}}{An edibble design.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the vertex.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-reactivate"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-reactivate}{}}}
\subsection{Method \code{reactivate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$reactivate(des, overwrite = c("graph", "anatomy", "recipe"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-deactivate"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-deactivate}{}}}
\subsection{Method \code{deactivate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$deactivate(delete = c("graph", "anatomy", "recipe"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_id_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_id_by_name}{}}}
\subsection{Method \code{fct_id_by_name()}}{
Get the id based on either the name of the factor node.
If none supplied then it will give all.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_id_by_name(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the vertex.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_id_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_id_by_class}{}}}
\subsection{Method \code{fct_id_by_class()}}{
Get all ids associated with a class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_id_by_class(class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_id_by_value"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_id_by_value}{}}}
\subsection{Method \code{lvl_id_by_value()}}{
Get the id based on name of level node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_id_by_value(value = NULL, fid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_names"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_names}{}}}
\subsection{Method \code{fct_names()}}{
Get the factor names based on id or class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_names(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_names"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_names}{}}}
\subsection{Method \code{lvl_names()}}{
Get the level names based on id or class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_names(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-append_fct_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-append_fct_nodes}{}}}
\subsection{Method \code{append_fct_nodes()}}{
Given node data, append the factor nodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$append_fct_nodes(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The nodes data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-append_lvl_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-append_lvl_nodes}{}}}
\subsection{Method \code{append_lvl_nodes()}}{
Given node data, append the level nodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$append_lvl_nodes(data, fid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The nodes data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-append_fct_edges"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-append_fct_edges}{}}}
\subsection{Method \code{append_fct_edges()}}{
Given edge data, append the factor edges
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$append_fct_edges(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The nodes data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-append_lvl_edges"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-append_lvl_edges}{}}}
\subsection{Method \code{append_lvl_edges()}}{
Given edge data, append the level edges
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$append_lvl_edges(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The nodes data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_class"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_class}{}}}
\subsection{Method \code{fct_class()}}{
Get the class of the vertex given the factor id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_class(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_class"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_class}{}}}
\subsection{Method \code{lvl_class()}}{
Get the class of the vertex given the level id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_class(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_child"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_child}{}}}
\subsection{Method \code{fct_child()}}{
Get the factor child ids. If \code{class} is
supplied then the child has to fit \code{class}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_child(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_child"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_child}{}}}
\subsection{Method \code{lvl_child()}}{
Get the level child ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_child(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_id_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_id_by_class}{}}}
\subsection{Method \code{lvl_id_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_id_by_class(class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_parent"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_parent}{}}}
\subsection{Method \code{fct_parent()}}{
Get the factor parent ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_parent(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_parent"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_parent}{}}}
\subsection{Method \code{lvl_parent()}}{
Get the level parent ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_parent(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_ancestor"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_ancestor}{}}}
\subsection{Method \code{fct_ancestor()}}{
Get the factor ancestor ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_ancestor(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-lvl_ancestor"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-lvl_ancestor}{}}}
\subsection{Method \code{lvl_ancestor()}}{
Get the level ancestor ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$lvl_ancestor(id = NULL, class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_levels"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_levels}{}}}
\subsection{Method \code{fct_levels()}}{
Get the levels for each factor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_levels(id = NULL, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the corresponding node.}

\item{\code{name}}{The name of the vertex.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-fct_exists"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-fct_exists}{}}}
\subsection{Method \code{fct_exists()}}{
One of \code{name}, \code{id} or \code{class} is defined to check if it exists.
If more than one of the arguments \code{name}, \code{id} and \code{class} are supplied, then
the intersection of it will be checked.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$fct_exists(name = NULL, id = NULL, class = NULL, abort = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the vertex.}

\item{\code{id}}{The id of the corresponding node.}

\item{\code{class}}{The class for the vertex/node.}

\item{\code{abort}}{A logical value to indicate whether to abort if it doesn't exist.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-trts_exists"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-trts_exists}{}}}
\subsection{Method \code{trts_exists()}}{
Check if treatment exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$trts_exists(abort = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abort}}{Whether to abort.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-units_exists"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-units_exists}{}}}
\subsection{Method \code{units_exists()}}{
Check if unit exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$units_exists(abort = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abort}}{Whether to abort.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-rcrds_exists"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-rcrds_exists}{}}}
\subsection{Method \code{rcrds_exists()}}{
Check if record exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$rcrds_exists(abort = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abort}}{Whether to abort.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-save_seed"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-save_seed}{}}}
\subsection{Method \code{save_seed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$save_seed(seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-get_history}{}}}
\subsection{Method \code{get_history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$get_history()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-get_seed"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-get_seed}{}}}
\subsection{Method \code{get_seed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$get_seed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-get_session_info"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-get_session_info}{}}}
\subsection{Method \code{get_session_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$get_session_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-get_edibble_version"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-get_edibble_version}{}}}
\subsection{Method \code{get_edibble_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$get_edibble_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-record_step"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-record_step}{}}}
\subsection{Method \code{record_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$record_step()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-record_history_external"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-record_history_external}{}}}
\subsection{Method \code{record_history_external()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$record_history_external(code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-record_history_internal"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-record_history_internal}{}}}
\subsection{Method \code{record_history_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$record_history_internal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-add_history_internal"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-add_history_internal}{}}}
\subsection{Method \code{add_history_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$add_history_internal(code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Provenance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Provenance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Provenance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
