% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{pull.edbl_nexus}
\alias{pull.edbl_nexus}
\title{Extract a single variable}
\usage{
pull.edbl_nexus(x, var = -1, ...)
}
\arguments{
\item{x}{\code{edbl_nexus} object}

\item{var}{variable name or position (see \link[dplyr]{pull})}

\item{name}{Optional vector used as name of extracted variable.}
}
\description{
This is similar to the \code{dplyr::pull} function but for the
\code{edbl_nexus} object which requires to be transformed to data frame first.
}
