% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise.R
\name{randomise_trts}
\alias{randomise_trts}
\title{Randomise treatments}
\usage{
randomise_trts(.design, ...)
}
\arguments{
\item{.design}{AN edibble graph object.}
}
\description{
Randomise treatments
}
\seealso{
Set the treatments by \code{\link[=set_trts]{set_trts()}}, set units by \code{\link[=set_units]{set_units()}},
and specify which treatment factors are applied to which units by using
\code{\link[=apply_trts]{apply_trts()}}.

Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{record_vars}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
