% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{render_table}
\alias{render_table}
\title{Serve edibble table}
\usage{
render_table(.design, ...)
}
\arguments{
\item{.design}{An \code{edbl_graph} object.}
}
\value{
An \code{edbl} data frame with columns defined by vertices and
rows displayed only if the vertices are connected and reconcile for output.
}
\description{
This converts an edibble graph object to a data frame called edibble.
}
