% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise.R
\name{replicabble}
\alias{replicabble}
\title{Make a replicabble}
\usage{
replicabble(.design, unit, trts)
}
\arguments{
\item{flist}{list of treatment variables with their levels}

\item{.replicate}{the number of units to allocate for every level.
It may be a vector but the length needs to match the order and
the number of combined levels of given trait variables.}
}
\description{
Make a replicabble
}
