% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_rcrd}
\alias{rescale_rcrd}
\title{Rescale a numerical vector}
\usage{
rescale_rcrd(x, lower = NA, upper = NA)
}
\arguments{
\item{x}{A numerical vector.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\description{
Similar to \code{\link[scales:rescale]{scales::rescale()}} but it has a different
behaviour when only upper or lower bound is given.
}
