% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{restart_design}
\alias{restart_design}
\title{Restart the edibble design}
\usage{
restart_design(.data)
}
\description{
This restarts the edibble design after initiating the design using
\code{\link[=edibble]{edibble()}}.
}
