% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rye_grass}
\alias{rye_grass}
\title{Rye-grass experiment}
\format{
An object of class \code{edbl_table} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 6 columns.
}
\source{
Bailey, Rosemary (2008) Design of Comparative Experiments.
}
\usage{
rye_grass
}
\description{
An experiment conducted to compare three different cultivars of rye-grass
(Cropper, Melba, Melle) in combination with four quantities of
nitrogen fertilizer (0 kg/ha, 80 kg/ha, 160 kg/ha, 240 kg/ha).

The experiment utilised two fields with each divided into three strips of
land. Each strip consisted of four plots.

\describe{
\item{strip}{A strip of land.}
\item{plot}{A plot within the strip.}
\item{cultivar}{The cultivars tested.}
\item{fertilizer}{The quantitites of nitrogen fertilizer used in kg/ha.}
\item{weight}{The total weight of dry mass after harvest.}
\item{perc}{The percentage of water-soluble carbohydrate in the crop.}
}
}
\seealso{
Other experimental data: 
\code{\link{lady_tasting_tea}},
\code{\link{skittles}}
}
\concept{experimental data}
\keyword{datasets}
