% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_levels}
\alias{sample_levels}
\title{Sample a subset of the levels for the given variable}
\usage{
sample_levels(.data, ..., n, prop, weight_by = NULL)
}
\arguments{
\item{.data}{An edibble table or edibble graph.}

\item{...}{The edibble variable(s) in \code{.data} to subset levels from.}

\item{n, prop}{Supply either \code{n}, the number of levels to sample, or \code{prop},
the proportion of levels to sample.}
}
\description{
Sample a subset of the levels for the given variable
}
