% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{scan_menu}
\alias{scan_menu}
\title{Find the short names of the named designs}
\usage{
scan_menu(pkgs = NULL)
}
\arguments{
\item{pkgs}{A character vector containing the package names to search
named designs from. By default it will search edibble and other packages loaded.}
}
\value{
A character vector of the short names of the named menu designs.
}
\description{
Find the short names of the named designs
}
\examples{
scan_menu()
}
