% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{select_units}
\alias{select_units}
\title{Select a subset of units from a cooked design}
\usage{
select_units(prep, ...)
}
\arguments{
\item{prep}{A cooked design.}

\item{...}{The units to select.}
}
\value{
An edibble design.
}
\description{
Select a subset of units from a cooked design
}
