% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_table.edbl_nexus}
\alias{serve_table.edbl_nexus}
\title{Serve edibble table}
\usage{
\method{serve_table}{edbl_nexus}(.nexus, ...)
}
\arguments{
\item{.nexus}{An \code{edbl_nexus} object.}
}
\value{
An \code{edbl} data frame with columns defined by vertices and
rows displayed only if the vertices are connected and reconcile for output.
}
\description{
This converts an edibble nexus object to a data frame called edibble.
}
