% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_table}
\alias{serve_table}
\title{Serve edibble table}
\usage{
serve_table(.design, use_labels = FALSE, ..., .record = TRUE)
}
\arguments{
\item{use_labels}{To show the labels instead of names.}

\item{...}{Strings of contexts or notes for the experiment. The context
maybe named. If you context it is wrong, consider writing the context
in an external file then write the reference to the file in context
instead. Input strings support inline markup that use \href{https://github.com/tidyverse/glue}{glue braces}
as well \href{https://cli.r-lib.org/articles/semantic-cli.html#inline-text-formatting}{cli markup formatter}.
The formatting is evaluated and stored in \code{edbl_design}.}

\item{.record}{A logical value. This indicates whether to record this
code step. The default is TRUE. It should remain TRUE unless this
function is used as a wrapper in other code.}
}
\value{
An \code{edbl} data frame with columns defined by vertices and
rows displayed only if the vertices are connected and reconcile for output.
}
\description{
This converts an edibble graph object to a data frame called edibble.
This function should be used when the design is in the final form
(or close to the final form). The table can only be formed when the
variables can be reconciled, otherwise it will be a data frame with
zero rows.
}
\seealso{
Other user-facing functions: 
\code{\link{allot_trts}()},
\code{\link{design-context}},
\code{\link{design}()},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()}
}
\concept{user-facing functions}
