% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{set_context}
\alias{set_context}
\title{Describe context related to experiment}
\usage{
set_context(.edibble, ...)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_df}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Strings of contexts or notes for the experiment. The context
maybe named. If you context it is wrong, consider writing the context
in an external file then write the reference to the file in context
instead. Input strings support inline markup that use \href{https://github.com/tidyverse/glue}{glue braces}
as well \href{https://cli.r-lib.org/articles/semantic-cli.html#inline-text-formatting}{cli markup formatter}.
The formatting is evaluated and stored in \code{EdibbleDesign}.}
}
\value{
Same as original except \code{EdibbleDesign} updated with context.
}
\description{
This is a function that stores a simple context about the experiment.
The context is printed when the return object is printed as well
as exported in the title sheet (see \code{\link[=export_design]{export_design()}}). If the context
already exists then it will be overwritten.
}
\examples{
start_design("COVID-19") \%>\%
  set_context(question = "Does Pfizer vaccine work?",
                 where = "Tested in lab",
                 "Context do not have to be named")
}
\seealso{
\code{\link[=sort_context]{sort_context()}} for sorting context alphabetically.
}
