% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{set_fcts}
\alias{set_fcts}
\title{Set edibble variables}
\usage{
set_fcts(
  .edibble,
  ...,
  .class = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  .code = NULL
)
}
\arguments{
\item{...}{Either a name-value pair or a series of the names.}

\item{.class}{A class for the variables.}

\item{.name_repair}{Same as the argument in \code{tibble::tibble()}.}
}
\description{
Adds variable and their level nodes in an edibble graph.
}
\seealso{
\code{\link[=set_units]{set_units()}} and \code{\link[=set_trts]{set_trts()}} for setting special types of nodes.
}
