% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{set_trts}
\alias{set_trts}
\title{Set the treatment variables}
\usage{
set_trts(.data, ...)
}
\arguments{
\item{.data}{An \code{edbl_graph} or \code{edbl_df} object.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}><\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Name-value pair.}
}
\description{
This function creates a special class, called \code{edbl_trt}, of edibble variables.
}
\section{Definition of \emph{treatment}}{

The word \emph{treatment} is sometimes used to refer to one of these variables.
When there are more than one treatment variables then this unfortunately
confuses whether treatment refers to the variable or the combination of
all treatment variables.

Treatment is the whole description of what is applied in an experiment.
}

\seealso{
See \code{\link[=set_units]{set_units()}} for setting units.
}
