% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{set_vars}
\alias{set_vars}
\title{Set edibble variables}
\usage{
set_vars(
  .edibble,
  ...,
  .class = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_df}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}><\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Name-value pair.}

\item{.name_repair}{Specify how to deal when there are duplicated name
entries. The repair follows the same argument in \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\description{
Adds variable and their level nodes in an edibble graph.
}
\seealso{
\code{\link[=set_units]{set_units()}} and \code{\link[=set_trts]{set_trts()}} for setting special types of nodes.
}
