% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_rcrds}
\alias{simulate_rcrds}
\title{Simulate records}
\usage{
simulate_rcrds(.data, ...)
}
\arguments{
\item{.data}{An edibble table or data frame}

\item{...}{A name-value pair.}
}
\description{
Simulate records
}
