% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{start_design}
\alias{start_design}
\title{Initiate the edibble design}
\usage{
start_design(name = NULL)
}
\arguments{
\item{name}{Optional name used as title for printing the design.}
}
\value{
An \code{EdibbleGraph} object which is a special type of \code{R6} class.
}
\description{
This function doesn't really do much besides create a new EdibbleDesign.
}
\examples{
start_design("My design")
}
\seealso{
Add variables to this design with \code{\link[=set_units]{set_units()}}, \code{\link[=set_trts]{set_trts()}}, and
\code{\link[=measure_units]{measure_units()}}.
}
