% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset-edbl-graph}
\alias{subset-edbl-graph}
\alias{subset.edbl_graph}
\alias{subset_vars}
\alias{subset_levels}
\title{Subset an edibble graph}
\usage{
\method{subset}{edbl_graph}(.graph, ..., .vtype = c("all", "var", "level"))

subset_vars(.graph)

subset_levels(.graph)
}
\arguments{
\item{.graph}{An edibble graph.}

\item{...}{Expressions that return a logical value.}

\item{.vtype}{The vertex type to subset. By default, all vertices are returned.
For \code{"var"}, only vertices that have attribute \code{vtype} as "var" is returned.}
}
\description{
The \code{subset} method subsets vertices of edibble graph based on vertex attributes.
The \code{subset_vars} and \code{subset_levels} subsets the edibble graph to
variable node or level nodes, respectively. This is essentially what we
refer to as high-level view or low-level view of the edibble graph respectively.
}
\examples{
graph <- get_edibble_graph(rye_grass)
subset(graph, vname \%in\% c("strip", "plot"))
subset_vars(graph)
susbet_levels(graph)
}
