% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset-edbl-nexus}
\alias{subset-edbl-nexus}
\alias{subset.edbl_nexus}
\alias{subset_vars}
\alias{subset_levels}
\title{Subset an edibble nexus}
\usage{
\method{subset}{edbl_nexus}(.nexus, ..., .vtype = c("all", "var", "level"))

subset_vars(.nexus)

subset_levels(.nexus)
}
\arguments{
\item{.nexus}{An edibble nexus.}

\item{...}{Expressions that return a logical value.}

\item{.vtype}{The vertex type to subset. By default, all vertices are returned.
For \code{"var"}, only vertices that have attribute \code{vtype} as "var" is returned.}
}
\description{
This function subsets vertices of edibble nexus based on vertex attributes.
}
