% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_units}
\alias{table_units}
\title{Count of child units per parent unit}
\usage{
table_units(.edibble, parent, child)
}
\arguments{
\item{.edibble}{An edibble design or graph}

\item{parent}{A string with the name of the parent.}

\item{child}{A string with the child name.}
}
\description{
Count of child units per parent unit
}
