% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{takeout}
\alias{takeout}
\title{Create a classical named experimental design}
\usage{
takeout(recipe = NULL, output = TRUE)
}
\arguments{
\item{recipe}{The short name of the classical named experimental design. See
under Details for the available named designs.}

\item{output}{A logical value to indicate whether all output should be
printed or not or a vector of character (e.g. \code{c("info", "code", "table")}) specifying which of the three
outputs should be printed. Default is TRUE.}
}
\description{
This function is defunct. Please use \code{takeout} instead.

The function \code{make_classical} generates a classical named experimental
design by supplying its short name and prints out, by default:
\itemize{
\item \code{info}: information about the named experimental design,
\item \code{code}: code to create the design using edibble, and
\item \code{table}: an edibble data frame for the generated design.
}

You can find the available short names with \code{find_classical_names()}.
}
\examples{
takeout(menu_crd(n = 50, t = 5))
# if you omit the design parameters then it will use the default
# (which may be random)
takeout(menu_crd())
# if you don't give any short names then it will generate a random one
takeout()
}
\seealso{
See \code{\link[=scan_menu]{scan_menu()}} for finding the short names of the
named experimental designs.
}
