% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{tbl_sum.edbl_df}
\alias{tbl_sum.edbl_df}
\title{Meta information print out for edibble}
\usage{
\method{tbl_sum}{edbl_df}(.data)
}
\arguments{
\item{.data}{An edibble data frame.}
}
\description{
Meta information print out for edibble
}
