% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{traits}
\alias{traits}
\title{Helper to set the traits for edibble variable}
\usage{
traits(
  labels = character(),
  levels = unique(labels),
  n = NULL,
  name = NULL,
  prefix = name,
  nlevels = n,
  rep = NULL,
  ...,
  class = NULL
)
}
\arguments{
\item{labels}{A character vector. If \code{levels} is supplied then the all
all elements should appear in \code{levels}.}

\item{levels}{A character vector with the name of the levels or
a named vector where the names are the name of the levels and the entries
are the number of replicate for the corresponding level.}

\item{n, nlevels}{The number of levels.}

\item{name}{The name of the object stored in the traits.}

\item{prefix}{The prefix to add the names of the levels.
Ignored if \code{levels} is used.}

\item{rep}{The number of available replicate for each level. If undefined,
the number of replicate is assumed to be infinite.}

\item{...}{Passed to \code{new_edibble_var}.}

\item{class}{The subclass name.}
}
\value{
Returns an \code{edbl_var} object.
}
\description{
This explicitly defines \code{edbl_var}.
}
\examples{
traits(n = 10, name = "plot")
traits(labels = c("vaccine", "control", "vaccine"),
       levels = c("vaccine", "control"))
}
