% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{traits_levels}
\alias{traits_levels}
\title{Helper function to create level names}
\usage{
traits_levels(
  prefix = "",
  size = NULL,
  from = NULL,
  to = NULL,
  include_leading_zero = edibble_labels_opt("leading_zero")
)
}
\description{
Helper function to create level names
}
