% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{trts_levels}
\alias{trts_levels}
\title{Unlike \code{vars_levels}, there is a check in place
that it is a treatment.}
\usage{
trts_levels(.data, vnames = NULL)
}
\description{
Unlike \code{vars_levels}, there is a check in place
that it is a treatment.
}
