% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise.R
\name{unit_to_trts}
\alias{unit_to_trts}
\title{Get the treatment variables for every unit}
\usage{
unit_to_trts(.design, unit, trts, type = c("var", "level"))
}
\value{
A named list where names are the unit variable name and
element is the character vector with the treatment factors applied
}
\description{
Get the treatment variables for every unit
}
