% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{utility-edibble-var}
\alias{utility-edibble-var}
\alias{as.character.edbl_fct}
\alias{as.integer.edbl_fct}
\alias{is_fct}
\alias{is_unit}
\alias{is_trt}
\alias{is_rcrd}
\title{Utility functions for edibble variable}
\usage{
\method{as.character}{edbl_fct}(x, ...)

\method{as.integer}{edbl_fct}(x, ...)

is_fct(x)

is_unit(x)

is_trt(x)

is_rcrd(x)
}
\arguments{
\item{x}{An \code{edbl_fct} object.}

\item{...}{Ignored.}
}
\value{
A character vector.
}
\description{
The S3 methods for \code{edbl_fct} objects have
the same expected output that of a factor.

Other functions are utility functions related to \code{edbl_fct} object.
}
