% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{var_ids}
\alias{var_ids}
\alias{trt_ids}
\alias{unit_ids}
\alias{rcrd_ids}
\alias{fct_id}
\alias{lvl_id}
\alias{fct_child}
\alias{lvl_child}
\alias{fct_parent}
\alias{lvl_parent}
\alias{fct_ancestor}
\alias{lvl_ancestor}
\title{Get the vertex IDs of the edibble design}
\usage{
var_ids(design, type = c("fct", "lvl"), vclass = NULL)

trt_ids(design, type = "fct")

unit_ids(design, type = "fct")

rcrd_ids(design, type = "fct")

fct_id(design, label = NULL)

lvl_id(design, label = NULL)

fct_child(design, id = NULL)

lvl_child(design, id = NULL)

fct_parent(design, id = NULL)

lvl_parent(design, id = NULL)

fct_ancestor(design, id = NULL)

lvl_ancestor(design, id = NULL)
}
\arguments{
\item{design}{An edibble design.}

\item{type}{A string value of either \code{"fct"} or \code{"lvl"}. Use \code{"fct"} for the
factor graph or \code{"lvl"} for the level graph.}

\item{vclass}{The class for the vertex/node.}
}
\description{
Set of operators to get the vertex/node IDs of an \code{edbl_design} object.
}
\seealso{
Other design manipulators: 
\code{\link{design_data}},
\code{\link{var_labels}()}
}
\concept{design manipulators}
