% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{var_labels}
\alias{var_labels}
\alias{trt_labels}
\alias{unit_labels}
\alias{rcrd_labels}
\alias{fct_label}
\alias{lvl_label}
\alias{fct_levels}
\title{Get the node labels of the edibble design}
\usage{
var_labels(design, vclass = NULL)

trt_labels(design)

unit_labels(design)

rcrd_labels(design)

fct_label(design, id = NULL)

lvl_label(design, id = NULL)

fct_levels(design, id = NULL, label = NULL)
}
\arguments{
\item{design}{An edibble design.}

\item{vclass}{The class for the vertex/node.}
}
\description{
Get the node labels of the edibble design
}
\seealso{
Other design manipulators: 
\code{\link{design_data}},
\code{\link{var_ids}()}
}
\concept{design manipulators}
