% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{var_names}
\alias{var_names}
\alias{trt_names}
\alias{unit_names}
\alias{rcrd_names}
\alias{fct_names}
\alias{lvl_names}
\alias{fct_levels}
\title{Get the node labels of the edibble design}
\usage{
var_names(design, vclass = NULL)

trt_names(design)

unit_names(design)

rcrd_names(design)

fct_names(design, id = NULL)

lvl_names(design, id = NULL)

fct_levels(design, id = NULL, name = NULL)
}
\arguments{
\item{design}{An edibble design.}

\item{vclass}{The class for the vertex/node.}
}
\description{
Get the node labels of the edibble design
}
\seealso{
Other design manipulators: 
\code{\link{design_data}},
\code{\link{var_ids}()}
}
\concept{design manipulators}
