% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{vgraph}
\alias{vgraph}
\title{Get the variable and level graph}
\usage{
vgraph(design, node_var = NULL, edge_var = NULL)
}
\arguments{
\item{design}{An edibble design}

\item{node_var}{The name of the column to select from the nodes}

\item{edge_var}{The name of the column to select from the edges
Ignored if \code{node_var} is defined instead.}
}
\description{
This function computes extra attributes such as
the the size and label names in the edges.
}
