% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{with_value}
\alias{with_value}
\title{Validation values}
\usage{
with_value(
  operator = c("=", "==", ">=", "<=", "<", ">", "!="),
  value = NULL,
  between = NULL,
  not_between = NULL
)
}
\arguments{
\item{operator}{Operator to apply.}

\item{value}{An optional value related to operator}

\item{between, not_between}{An optional numerical vector of size two where the
first entry is the minimum value and the second entry is the maximum value.
For \code{between}, the value is valid if within the range of minimum and maximum
value inclusive. For \code{not_between}, the value must lie outside of these values.}
}
\value{
A list with two elements \code{operator} and \code{value}.
}
\description{
This creates a list that is used later for creating data validation rules
when the data is exported.
}
