% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{add_context}
\alias{add_context}
\title{Describe context related to experiment}
\usage{
add_context(.data, ..., .overwrite = TRUE)
}
\arguments{
\item{.data}{An edibble graph.}

\item{...}{Name-value pairs.}

\item{.overwrite}{If the context}
}
\description{
This is just a function that stores a simple context. If the
context already exists then it will be overwritten.
}
\examples{
initiate_design("COVID-19") \%>\%
  describe_context(question = "Does Pfizer vaccine work?",
                   where = "Tested in lab")

}
