% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{add_edibble_vertex}
\alias{add_edibble_vertex}
\alias{add_edibble_vertex.default}
\alias{add_edibble_vertex.lkbl_graph}
\title{Add an edibble vertex}
\usage{
add_edibble_vertex(.data, ...)

\method{add_edibble_vertex}{default}(value, name, design = NULL, attr)

\method{add_edibble_vertex}{lkbl_graph}(.lgraph, name, design = NULL, attr)
}
\arguments{
\item{value}{A value that may be a single number, unnamed vector,
a one sided formula or linkabble.}

\item{name}{Name of the edibble variable as string.}

\item{design}{An edibble graph with levels converted as list used
for evaluation in any expressions.}
}
\value{
Returns an evaluated expression.
}
\description{
In edibble, a short hand is used to assign new edibble variables/nodes.
This function converts the short hand to make the minimal edibble node
and returns an edibble graph with that additional node.
}
