% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{allocate_trts}
\alias{allocate_trts}
\title{Define which unit to apply treatment}
\usage{
allocate_trts(.design, ...)
}
\arguments{
\item{...}{One-sided or two-sided formula. If the input is a one-sided formula
then the whole treatment is applied to the specified unit.}

\item{.data}{An \code{edbl_graph} object.}

\item{class}{A sub-class. This is meant so that it can invoke the method
\code{randomise_trts.class}.}
}
\description{
Define which unit to apply treatment
}
\seealso{
Other user-facing functions: 
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{record_vars}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
