% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{allocate_trts}
\alias{allocate_trts}
\title{Define which unit to apply treatment}
\usage{
allocate_trts(.data, ..., class = NULL)
}
\arguments{
\item{.data}{An \code{edbl_graph} object.}

\item{...}{One-sided or two-sided formula. If the input is a one-sided formula
then the whole treatment is applied to the specified unit.}

\item{class}{A sub-class. This is meant so that it can invoke the method
\code{randomise_trts.class}.}
}
\description{
Define which unit to apply treatment
}
