% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allot.R
\name{allot_table}
\alias{allot_table}
\title{A shorthand for allot, assign and serve}
\usage{
allot_table(
  .design,
  ...,
  order = "random",
  seed = NULL,
  constrain = nesting(.design)
)
}
\arguments{
\item{.design}{An edibble design which should have units, treatments and allotment defined.}

\item{...}{
  Arguments passed on to \code{\link[=allot_trts]{allot_trts}}
  \describe{
    \item{\code{.record}}{A logical value. This indicates whether to record this
code step. The default is TRUE. It should remain TRUE unless this
function is used as a wrapper in other code.}
  }}

\item{order}{A character vector signifying the apportion of treatments to units.
The value should be either "random", "systematic" or "systematic-random".
"random" allocates the treatment randomly to units based on specified allotment with restrictions
implied by unit structure.
"systematic" allocates the treatment in a systematic order to units.
"systematic-random" allocates the treatment in a systematic order to units but
where it is not possible to divide treatments equally (as the number of units are not divisible
by the number of levels of the treatment factor), then the extras are chosen randomly.}

\item{seed}{A scalar value used to set the seed so that the result is reproducible.}

\item{constrain}{The nesting structure for units.}
}
\description{
A shorthand for allot, assign and serve
}
