% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allot.R
\name{allot_table}
\alias{allot_table}
\title{Allot treatments to units and serve table}
\usage{
allot_table(
  .edibble,
  ...,
  order = "random",
  seed = NULL,
  constrain = nesting_structure(.edibble)
)
}
\description{
This function is a short hand that combines \code{allot_trts()}, \code{assign_trts()}
and \code{serve_table()}.
}
