% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{allot_trts}
\alias{allot_trts}
\title{Define the possible allocation of treatments to units}
\usage{
allot_trts(design, ...)
}
\arguments{
\item{...}{One-sided or two-sided formula. If the input is a one-sided formula
then the whole treatment is applied to the specified unit.}

\item{.data}{An \code{edbl_graph} object.}
}
\description{
This function adds the edges between variable nodes to
specify the mapping of units to treatment. This function
does not actually assign specific treatment levels onto actual units.
}
\examples{
start_design() \%>\%
  set_units(block = 10,
            plot = nested_in(block, 3)) \%>\%
  set_trts(treat = c("A", "B", "C"),
           pest = c("a", "b")) \%>\%
  allot_trts(treat ~ plot,
                pest ~ block)


}
\seealso{
Other user-facing functions: 
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
