% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allot.R
\name{allot_trts}
\alias{allot_trts}
\title{Define the possible allocation of treatments to units}
\usage{
allot_trts(.design, ..., .record = TRUE)
}
\arguments{
\item{...}{One-sided or two-sided formula. If the input is a one-sided formula
then the whole treatment is applied to the specified unit.}

\item{.record}{A logical value. This indicates whether to record this
code step. The default is TRUE. It should remain TRUE unless this
function is used as a wrapper in other code.}
}
\value{
Return an edibble design.
}
\description{
This function adds the edges between variable nodes to
specify the mapping of units to treatment. This function
does not actually assign specific treatment levels onto actual units.
}
\examples{
start_design() \%>\%
  set_units(block = 10,
            plot = nested_in(block, 3)) \%>\%
  set_trts(treat = c("A", "B", "C"),
           pest = c("a", "b")) \%>\%
  allot_trts(treat ~ plot,
                pest ~ block)

}
\seealso{
Other user-facing functions: 
\code{\link{design-context}},
\code{\link{design}()},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()}
}
\concept{user-facing functions}
