% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{classics}
\alias{classics}
\alias{nclassics}
\title{A list of classical experimental designs}
\format{
A list of some classical experimental designs as either \code{edbl_nexus}
or \code{edbl} object. The designs included are:
\describe{
\item{CRD} Completely randomised design.
\item{RCBD} Randomised complete block design.
\item{split} Split plot design.
}
}
\usage{
nclassics
}
\description{
A list of classical experimental designs
}
\keyword{datasets}
