% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{design-helpers}
\alias{design-helpers}
\alias{is_edibble_design}
\alias{is_named_design}
\alias{is_edibble_df}
\alias{is_edibble_table}
\alias{is_edibble_graph}
\alias{is_edibble}
\alias{get_edibble_design}
\alias{get_edibble_table}
\alias{get_edibble_graph}
\title{Test and get edibble objects}
\usage{
is_edibble_design(x)

is_named_design(x)

is_edibble_df(x)

is_edibble_table(x)

is_edibble_graph(x)

is_edibble(x)

get_edibble_design(x)

get_edibble_table(x)

get_edibble_graph(x)
}
\arguments{
\item{x}{An object.}
}
\description{
The \code{is} functions tests if an object (or an object in its attribute)
inherits particular class and returns \code{TRUE} if it does, otherwise \code{FALSE}.
\itemize{
\item \code{is_edibble_design} checks if it inherits \code{EdibbleDesign}.
\item \code{is_edibble_graph} checks if it inherits \code{edbl_graph}.
\item \code{is_edibble_df}, \code{is_edibble_table} checks if it inherits \code{edbl_df}
\item \code{is_edibble} checks if the object contains \code{EdibbleDesign}.
The search is quite simple, it checks if
the object is \code{EdibbleDesign}, failing that it looks to see if the
attribute "design" of the object is \code{EdibbleDesign}.
\item \code{is_named_design} check if it inherits \code{NamedDesign}.
}

The \code{get} functions extracts the requested edibble component (table, graph,
or design) from the object if possible.
\itemize{
\item \code{get_edibble_design} tries to get \code{EdibbleDesign}.
\item \code{get_edibble_table} tries to get \code{edbl_df} with no design attribute.
\item \code{get_edibble_graph} tries to get \code{edbl_graph}.
}
}
\examples{
is_edibble_df(mtcars) # FALSE
is_named_design(code_classical_rcbd(.quiet = TRUE)) # TRUE
}
