% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{design_data}
\alias{design_data}
\alias{fct_nodes}
\alias{fct_edges}
\alias{lvl_nodes}
\alias{lvl_edges}
\title{Get the node or edge data from an edibble design}
\usage{
fct_nodes(x)

fct_edges(x)

lvl_nodes(x)

lvl_edges(x)
}
\arguments{
\item{x}{An edibble object.}
}
\description{
Get the node or edge data from an edibble design
}
\concept{design manipulators}
