% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble-package.R
\docType{package}
\name{edibble-package}
\alias{edibble-package}
\alias{_PACKAGE}
\title{edibble: Designing Comparative Experiments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system to facilitate designing comparative experiments using the grammar of experimental designs \url{https://emitanaka.org/edibble-book/}. An experimental design is treated as an intermediate, mutable object that is built progressively by fundamental experimental components like units, treatments, and their relation.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{(WIP)}
}
\section{Website}{

\itemize{
\item The website for the package is at \url{https://edibble.emitanaka.org}
}
\itemize{
\item Discussion is at \url{https://github.com/emitanaka/edibble/discussions}
}
}

\section{Package options}{

The following options are used for changing the default view for the print
out of edibble design or edibble graph.
\itemize{
\item \code{edibble.tree.decorate.trts}
\item \code{edibble.tree.decorate.units}
\item \code{edibble.tree.decorate.rcrd}
\item \code{edibble.tree.decorate.levels}
\item \code{edibble.tree.decorate.main}
}

TODO
}

\seealso{
Useful links:
\itemize{
  \item \url{https://edibble.emitanaka.org/}
  \item \url{https://github.com/emitanaka/edibble}
  \item Report bugs at \url{https://github.com/emitanaka/edibble/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
