% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{edibble}
\alias{edibble}
\alias{as_edibble}
\title{Make input edibble}
\usage{
edibble(.data, name = NULL, ..., units = NULL, trts = NULL)

as_edibble(.data, ...)
}
\arguments{
\item{.data}{A named list of variables or data frame.}

\item{...}{Passed to \code{new_tibble}.}

\item{units}{A character vector specifying variables names in \code{.data}
which are the unit variables.}

\item{trts}{A character vector specifying variables names in \code{.data}
which are the treatment variables.}
}
\description{
\strong{(WIP)} If variables are already defined as external data then you can import this
data as edibble.
}
\seealso{
See \code{\link[=start_design]{start_design()}} for initiating design from scratch.
}
