% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{examine_process}
\alias{examine_process}
\alias{examine_process_values}
\title{Examine the simulation process}
\usage{
examine_process(data, process = NULL)

examine_process_values(data, process = NULL)
}
\arguments{
\item{data}{An edibble data frame.}

\item{process}{The process name. Typically the name of the process. If unknown,
leave this empty.}
}
\description{
Examine the simulation process
}
