% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{examine_recipe}
\alias{examine_recipe}
\title{Check the recipe code}
\usage{
examine_recipe(x, ...)
}
\arguments{
\item{x}{An edibble design, edibble, or takeout object.}

\item{...}{Not used.}
}
\description{
Check the recipe code
}
