% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{expect-vars}
\alias{expect-vars}
\alias{to_be_numeric}
\alias{to_be_integer}
\alias{to_be_date}
\alias{to_be_time}
\alias{to_be_character}
\alias{to_be_selected}
\title{Expected type of entry}
\usage{
to_be_numeric(range)

to_be_integer(range)

to_be_date(range)

to_be_time(range)

to_be_character(length)

to_be_selected(from)
}
\arguments{
\item{range, length}{A named list with two elements: "operator" and "value" as
provided by helper \code{as_value()} that gives the possible range of values
that the expected type can take.}

\item{value}{A vector of possible values for entry.}
}
\description{
Expected type of entry
}
