% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_design}
\alias{export_design}
\title{Export the design to xlsx}
\usage{
export_design(
  .data,
  file,
  author = NULL,
  date = Sys.Date(),
  overwrite = FALSE,
  subject = NULL,
  category = NULL,
  table_style = "TableStyleMedium9"
)
}
\arguments{
\item{.data}{An edibble table to export.}

\item{file}{File, including the path, to export the data to.}

\item{author}{(Optional) name of the author in character. A vector of character is supported
for where there are multiple authors.}

\item{date}{The date to be inserted in header (defaults to today).}

\item{overwrite}{A logical indicating whether to overwrite existing file or not.}

\item{subject}{The subject of the workbook (optional).}

\item{category}{The category of the workbook (optional).}

\item{table_style}{The table style to apply to the exported data (default: "TableStyleMedium9").}
}
\value{
The input data object.
}
\description{
This function is designed to export the design made using edibble to an
external xlsx file.
}
\seealso{
Other user-facing functions: 
\code{\link{allot_trts}()},
\code{\link{allot_units}()},
\code{\link{design}()},
\code{\link{expect_rcrds}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()}
}
\concept{user-facing functions}
