% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_design}
\alias{export_design}
\title{Export the design to xlsx}
\usage{
export_design(.data, file, author, date = Sys.Date(), overwrite = FALSE)
}
\arguments{
\item{.data}{An edibble data frame or design.}

\item{file}{File, including the path, to export the data to.}

\item{author}{Name of the author in character. A vector of character is supported
for where there are multiple authors.}

\item{date}{The date to be inserted in header.}

\item{overwrite}{A logical indicating whether to overwrite exisitng file or not.}
}
\description{
This function is designed to export the design made using edibble to an
external xlsx file.
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{randomise_trts}()},
\code{\link{record_vars}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
