% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{fct_attrs}
\alias{fct_attrs}
\title{Setting the traits of factors}
\usage{
fct_attrs(.levels = NULL, ...)
}
\arguments{
\item{.levels}{An \code{edbl_lvls} object that should contain information about the levels
in the factor.}

\item{...}{A name-value pair of attributes. The value must be a scalar and
attributed to the whole factor (not individual levels).
The values are added as attributes to the output object.}
}
\value{
An \code{edbl_lvls} object.
}
\description{
This function is used to set characteristics of the factors.
}
\examples{
fct_attrs(levels = c("A", "B"))
}
\seealso{
lvl_traits
}
