% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{is_edibble}
\alias{is_edibble}
\alias{not_edibble}
\title{Test if the object is an edibble data frame}
\usage{
is_edibble(x)

not_edibble(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits the \code{edbl_df} class.
}
\description{
A simple function that return \code{TRUE} if the object inherits
the \code{edbl_df} class, \code{FALSE} otherwise.
}
\seealso{
See \code{\link[=is_edibble_graph]{is_edibble_graph()}} for testing if the object is \code{edbl_graph}.
}
