% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{measure_response}
\alias{measure_response}
\title{Measure response of given units}
\usage{
measure_response(
  .nexus,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.nexus}{An \code{edbl_nexus} object}

\item{...}{Name-value pair.}
}
\description{
This function creates new nodes to edibble nexus with the name
corresponding to the intended response that will be measured.
}
