% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{menu_crd}
\alias{menu_crd}
\title{Completely randomised design}
\usage{
menu_crd(
  t = random_integer_small(),
  n = random_integer_medium(min = t),
  r = NULL,
  seed = random_seed_number()
)
}
\arguments{
\item{t}{The number of treatment levels}

\item{n}{The number of experimental units}

\item{r}{(Optional) The number of replicates.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Completely randomised design
}
\seealso{
Other named-designs: 
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_split}()},
\code{\link{menu_youden}()}
}
\concept{named-designs}
