% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_rcbd}
\alias{menu_rcbd}
\title{Prepare a randomised complete block design}
\usage{
menu_rcbd(
  t = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number()
)
}
\arguments{
\item{t}{The number of treatments.}

\item{r}{The number of replicate blocks.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare a randomised complete block design
}
\seealso{
Other recipe-designs: 
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_split}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
