% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_split}
\alias{menu_split}
\title{Split-unit design}
\usage{
menu_split(
  t1 = random_integer_small(),
  t2 = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number()
)
}
\arguments{
\item{t1}{The number of treatment levels for the main plots.}

\item{t2}{The number of treatment levels for the subplots.}

\item{r}{The number of replicate blocks.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Originally referred to as split-plot design when it was first used.
}
\seealso{
Other recipe-designs: 
\code{\link{menu_crd}()},
\code{\link{menu_factorial}()},
\code{\link{menu_graeco}()},
\code{\link{menu_hyper_graeco}()},
\code{\link{menu_lsd}()},
\code{\link{menu_rcbd}()},
\code{\link{menu_strip}()},
\code{\link{menu_youden}()}
}
\concept{recipe-designs}
