% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names-ed-vars}
\alias{names-ed-vars}
\alias{names_units}
\alias{names_trts}
\alias{names_resp}
\alias{names_vars}
\alias{names.edbl_nexus}
\alias{names<-.edbl_nexus}
\title{Names of variables}
\usage{
names_units(x)

names_trts(x)

names_resp(x)

names_vars(x, class = NULL)

\method{names}{edbl_nexus}(x)

\method{names}{edbl_nexus}(x) <- value
}
\arguments{
\item{x}{An edibble or edibble nexus.}

\item{class}{A string.}
}
\value{
A character vector of the names of the variables with the given class.
}
\description{
Functions to get or set the names of variables (with particular class) from
edibble or edibble nexus object.
}
\examples{
names_units(nclassics$split)
names_trts(nclassics$split)
}
