% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise.R
\name{names_to_vnames}
\alias{names_to_vnames}
\title{Get the variable names given the vertex name of edibble graph}
\usage{
names_to_vnames(.design, names)
}
\arguments{
\item{.design}{An edibble graph.}

\item{names}{A vector of character with vertex names.}
}
\value{
A character vector of the variable name.
}
\description{
Get the variable names given the vertex name of edibble graph
}
