% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nested_in}
\alias{nested_in}
\title{Specify the nesting structure for units}
\usage{
nested_in(.unit, ..., .vname)
}
\arguments{
\item{.unit}{The name of the parent unit to nest under.}

\item{...}{a single number OR a sequence of two-sided formula where the
left-hand side corresponds to the name of the level (or the level number) of \code{.unit}
and the right-hand side is an integer specifying the number of levels nested under the
corresponding levels.}

\item{.vname}{the child variable name}
}
\value{
Returns an isolated graph
}
\description{
Specify the nesting structure for units
}
\seealso{
See \code{\link[=set_units]{set_units()}} for examples of how to use this.
}
