% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nested_in}
\alias{nested_in}
\title{Specify the nesting structure for units}
\usage{
nested_in(.var, ..., .vname)
}
\arguments{
\item{...}{a single number OR a sequence of two-sided formula where the left-hand side corresponds to the level of \code{unit}
and the right-hand side is the levels of the new unit OR the number of levels}

\item{.vname}{an optional variable name}

\item{x}{A vector with the levels of the list with the name corresponding to
the name of the unit to nest under.}
}
\value{
Returns an isolated graph <- <- <- <-
}
\description{
Specify the nesting structure for units
}
