% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{new_edibble}
\alias{new_edibble}
\alias{as_edibble}
\alias{edibble}
\title{An edibble table constructor}
\usage{
new_edibble(.data, ..., design = NULL, class = NULL)

as_edibble(.data, ...)

edibble(
  .data,
  name = NULL,
  .record = TRUE,
  seed = NULL,
  kitchen = Kitchen,
  ...
)
}
\arguments{
\item{.data}{data frame or list of the same size.}

\item{...}{Passed to \code{new_tibble}.}

\item{class}{Subclasses for edibble table. The default is NULL.}

\item{graph}{An edibble graph object.}
}
\description{
This helps to construct a new edibble table which is a special type
of tibble.
}
