% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{new_edibble}
\alias{new_edibble}
\alias{as_edibble}
\title{An edibble table constructor}
\usage{
new_edibble(.data, ..., .design = NULL, .class = NULL)

as_edibble(.data, ...)
}
\arguments{
\item{.data}{data frame or list of the same size.}

\item{...}{Passed to \code{new_tibble}.}

\item{.design}{An edibble graph object.}

\item{.class}{Subclasses for edibble table. The default is NULL.}

\item{units}{The data columns that are units.}

\item{trts}{The data columns that are treatments.}
}
\value{
An edibble table.
}
\description{
This helps to construct a new edibble table which is a special type
of tibble.
}
