% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edibble.R
\name{new_edibble}
\alias{new_edibble}
\title{An edibble constructor}
\usage{
new_edibble(.data, ..., graph = NULL, class = NULL)
}
\arguments{
\item{.data}{data frame or list of the same size.}

\item{...}{Passed to \code{new_tibble}.}

\item{graph}{An edibble graph object.}

\item{class}{Subclasses for edibble. The default is NULL.}
}
\description{

}
