% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edbl_design}
\alias{plot.edbl_design}
\alias{plot.edbl_table}
\title{Interactive plot of the edibble design}
\usage{
\method{plot}{edbl_design}(
  .design,
  which = c("factors", "levels"),
  width = "100\%",
  height = NULL,
  seed = 1,
  title = NULL,
  view = c("show-buttons", "hide-buttons", "static"),
  ...
)

\method{plot}{edbl_table}(.table, ...)
}
\arguments{
\item{.design}{An edibble design.}

\item{which}{A string of either "factors" or "levels".}

\item{width, height}{The width and height of the plot.}

\item{seed}{A seed number so same plot is always generated.}

\item{title}{The title of the plot. By default it uses the name from the \code{.design} object.}

\item{view}{A string of either "show-buttons" (default), "hide-buttons", "static"}
}
\description{
Interactive plot of the edibble design
}
