% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_crd}
\alias{prep_classical_crd}
\title{Completely randomised design}
\usage{
prep_classical_crd(t = 1 + sample(10, 1), n = t + sample(100, 1), r = n/t)
}
\arguments{
\item{t}{The number of treatment levels}

\item{n}{The number of experimental units}

\item{r}{(Optional) The number of replicates.}
}
\description{
Completely randomised design
}
