% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_crd}
\alias{prep_classical_crd}
\title{Completely randomised design}
\usage{
prep_classical_crd(
  t = random_integer_small(),
  n = random_integer_medium(min = t),
  r = n/t,
  seed = random_seed_number()
)
}
\arguments{
\item{t}{The number of treatment levels}

\item{n}{The number of experimental units}

\item{r}{(Optional) The number of replicates.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Completely randomised design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_lsd}()},
\code{\link{prep_classical_rcbd}()},
\code{\link{prep_classical_split}()}
}
\concept{named-designs}
