% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classics.R
\name{prep_classical_split}
\alias{prep_classical_split}
\title{Prepare classical split plot design}
\usage{
prep_classical_split(
  t1 = random_integer_small(),
  t2 = random_integer_small(),
  r = random_integer_small(),
  seed = random_seed_number()
)
}
\arguments{
\item{t1}{The number of treatment levels for the main plots.}

\item{t2}{The number of treatment levels for the subplots.}

\item{r}{The number of replicate blocks.}

\item{seed}{A scalar value for computational reproducibility.}
}
\description{
Prepare classical split plot design
}
\seealso{
Other named-designs: 
\code{\link{prep_classical_crd}()},
\code{\link{prep_classical_factorial}()},
\code{\link{prep_classical_lsd}()},
\code{\link{prep_classical_rcbd}()}
}
\concept{named-designs}
