% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise.R
\name{randomise_trts}
\alias{randomise_trts}
\alias{randomise_trts.EdibbleDesign}
\title{Randomise treatments}
\usage{
randomise_trts(.edibble, ...)

\method{randomise_trts}{EdibbleDesign}(.edibble)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_table}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Currently unused.}
}
\description{
Randomise treatments
}
\seealso{
Set the treatments by \code{\link[=set_trts]{set_trts()}}, set units by \code{\link[=set_units]{set_units()}},
and specify which treatment factors are applied to which units by using
\code{\link[=allocate_trts]{allocate_trts()}}.

Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
