% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{record_vars}
\alias{record_vars}
\title{Measure response of given units}
\usage{
record_vars(
  .design,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.design}{An \code{EdibbleDesign} object}

\item{...}{Name-value pair. The name should correspond to the name of the
unit defined in \code{set_units}. The value should be a vector of new variables
names.}
}
\description{
This function creates new nodes to edibble graph with the name
corresponding to the intended response that will be measured.
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
