% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_table}
\alias{serve_table}
\title{Serve edibble table}
\usage{
serve_table(.design, ...)
}
\arguments{
\item{.design}{An \code{edbl_graph} object.}

\item{...}{Ignored.}
}
\value{
An \code{edbl} data frame with columns defined by vertices and
rows displayed only if the vertices are connected and reconcile for output.
}
\description{
This converts an edibble graph object to a data frame called edibble.
This function should be used when the design is in the final form
(or close to the final form). The table can only be formed when the
variables can be reconciled, otherwise it will be a data frame with
zero rows.
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{record_vars}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
