% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{set_rcrds}
\alias{set_rcrds}
\title{Set records for given unit}
\usage{
set_rcrds(
  .edibble,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_table}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Name-value pair. The name should correspond to the name of the
unit defined in \code{set_units}. The value should be a vector of new variables
names.}
}
\description{
This function creates new nodes to edibble graph with the name
corresponding to either the intended response that will be measured or
a variable to be recorded.
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
