% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trts.R
\name{set_trts}
\alias{set_trts}
\title{Set the treatment variables}
\usage{
set_trts(
  .design,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{...}{Either a name-value pair or a series of the names.}

\item{.name_repair}{Same as the argument in \code{tibble::tibble()}.}
}
\description{
This function add a special class, called \code{edbl_trt}, of edibble variables.
}
\section{Definition of \emph{treatment}}{

The word \emph{treatment} is sometimes used to refer to one of these variables.
When there are more than one treatment variables then this unfortunately
confuses whether treatment refers to the variable or the combination of
all treatment variables.

Treatment is the whole description of what is applied in an experiment.
}

\examples{

start_design() \%>\%
  set_trts(pesticide = c("A", "B", "C"),
           dosage = c(0, 10, 20, 30, 40))

# you can set treatments to existing edibble
lady_tasting_tea \%>\%
  edibble() \%>\%
  set_trts(first)
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_units}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
