% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{set_units}
\alias{set_units}
\title{Set units used in experiment}
\usage{
set_units(.data, ...)
}
\arguments{
\item{.data}{An \code{edbl_graph} or \code{edbl_df} object.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}><\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Name-value pair.}
}
\description{
This function creates new edibble variables of class \code{edbl_unit}.
}
\section{Definition of \emph{unit}}{

A \emph{unit}, much like \emph{factor}, is an over-used word but due to lack of a
better word, edibble uses the word "unit" to refer to any entity, physical
or otherwise, that pertain to the experiment. This function doen't
explicitly distinguish between experimental or observational units,
nor is a unit limited to these type of units.
A unit in edibble can be a blocking factor or discrete time unit.
}

\section{Limitations}{

Currently a unit should only have a discrete set of levels.
}

\examples{
initiate_design() \%>\%
  set_units(block = 3)
}
