% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{set_units}
\alias{set_units}
\title{Set units used in experiment}
\usage{
set_units(
  .edibble,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_table}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Either a name-value pair or a series of the names}
}
\description{
This function sets new edibble variables of class \code{edbl_unit}. More
specifically, this means that new nodes are added to the \code{edbl_graph}.
}
\section{Definition of \emph{unit}}{

A \emph{unit}, much like \emph{factor}, is an over-used word but due to lack of a
better word, edibble uses the word "unit" to refer to any entity, physical
or otherwise, that pertain to the experiment. This function doen't
explicitly distinguish between experimental or observational units,
nor is a unit limited to these type of units.
A unit in edibble can be a blocking factor or even a discrete time unit.
}

\section{Limitations}{

Currently a unit should only have a discrete set of levels and
you need to know the number of levels prior to setting the units.
}

\examples{
# 30 rats
start_design() \%>\%
  set_units(rat = 30) \%>\%
  serve_table()

# 4 girls named "Anna", "Betty", "Carol", "Diana"
start_design() \%>\%
  set_units(girl = c("Anna", "Betty", "Carol", "Diana")) \%>\%
  serve_table()

# 3 companies, with 10 boxes each
start_design() \%>\%
  set_units(company = c("A", "B", "C"),
                box = nested_in(company, 10))

}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{record_vars}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
