% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{set_units}
\alias{set_units}
\title{Set units used in experiment}
\usage{
set_units(
  .edibble,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.edibble}{An edibble design (\code{EdibbleDesign}), an edibble data frame (\code{edbl_df}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}><\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Name-value pair.}
}
\description{
This function creates new edibble variables of class \code{edbl_unit}.
}
\section{Definition of \emph{unit}}{

A \emph{unit}, much like \emph{factor}, is an over-used word but due to lack of a
better word, edibble uses the word "unit" to refer to any entity, physical
or otherwise, that pertain to the experiment. This function doen't
explicitly distinguish between experimental or observational units,
nor is a unit limited to these type of units.
A unit in edibble can be a blocking factor or discrete time unit.
}

\section{Limitations}{

Currently a unit should only have a discrete set of levels.
}

\examples{
initiate_design() \%>\%
  set_units(block = 3)
}
\seealso{
Other user-facing functions: 
\code{\link{allocate_trts}()},
\code{\link{design-context}},
\code{\link{expect_vars}()},
\code{\link{export_design}()},
\code{\link{randomise_trts}()},
\code{\link{record_vars}()},
\code{\link{serve_table}()},
\code{\link{set_trts}()},
\code{\link{start_design}()}
}
\concept{user-facing functions}
