% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{set_vars}
\alias{set_vars}
\title{Set edibble variables}
\usage{
set_vars(
  .edibble,
  ...,
  .class = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.edibble}{An edibble design (\code{edbl_design}), an edibble data frame (\code{edbl_table}) or an
object that contains the edibble data frame in the attribute
\code{design}.}

\item{...}{Either a name-value pair or a series of the names.}

\item{.class}{A class for the variables.}

\item{.name_repair}{Same as the argument in \code{tibble::tibble()}.}
}
\description{
Adds variable and their level nodes in an edibble graph.
}
\seealso{
\code{\link[=set_units]{set_units()}} and \code{\link[=set_trts]{set_trts()}} for setting special types of nodes.
}
