% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_rcrds}
\alias{simulate_rcrds}
\title{Simulate records}
\usage{
simulate_rcrds(.data, ..., .seed = NULL)
}
\arguments{
\item{.data}{An edibble data}

\item{...}{A name-value pair where the name should correspond to the names
used in the \code{\link[=simulate_process]{simulate_process()}}. The value should be returned from calling
\code{\link[=with_params]{with_params()}}.}

\item{.seed}{An optional seed value.}
}
\description{
Simulate records
}
