% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{start_design}
\alias{start_design}
\title{Start the edibble design}
\usage{
start_design(name = NULL)
}
\arguments{
\item{name}{Optional name used as title for printing the design.}
}
\value{
An empty \code{edbl_design} object.
}
\description{
This function doesn't really do much besides create a new edibble design object.
}
\examples{
start_design("My design")
}
\seealso{
Add variables to this design with \code{\link[=set_units]{set_units()}}, \code{\link[=set_trts]{set_trts()}}, and
\code{\link[=set_rcrds]{set_rcrds()}}.

Other user-facing functions: 
\code{\link{allot_trts}()},
\code{\link{design-context}},
\code{\link{expect_rcrds}()},
\code{\link{export_design}()},
\code{\link{serve_table}()},
\code{\link{set_rcrds}()},
\code{\link{set_trts}()},
\code{\link{set_units}()}
}
\concept{user-facing functions}
