% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset-edbl-graph}
\alias{subset-edbl-graph}
\alias{subset.edbl_graph}
\alias{subset_vars}
\alias{subset_levels}
\title{Subset an edibble graph}
\usage{
\method{subset}{edbl_graph}(.graph, ..., .vtype = c("all", "var", "level"))

subset_vars(.graph)

subset_levels(.graph)
}
\arguments{
\item{.graph}{An edibble graph.}

\item{...}{Expressions that return a logical value.}

\item{.vtype}{The vertex type to subset. By default, all vertices are returned.
For \code{"var"}, only vertices that have attribute \code{vtype} as "var" is returned.}
}
\description{
This function subsets vertices of edibble graph based on vertex attributes.
}
