% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{takeout}
\alias{takeout}
\title{Create a named experimental design}
\usage{
takeout(recipe = NULL, output = FALSE)
}
\arguments{
\item{recipe}{A named design object. This should be typically generated from a
function with prefix \code{menu_}. If nothing is supplied, it will randomly select one.}

\item{output}{A logical value to indicate whether all output should be
printed or not or a vector of character (e.g. \code{c("info", "code", "table")}) specifying which of the three
outputs should be printed. Default is TRUE.}
}
\description{
This function generates a named experimental
design by supplying the selected menu named design and prints out, by default:
\itemize{
\item \code{info}: information about the named experimental design, and
\item \code{code}: code to create the design using the fundamental system.
}

You can find the available recipes with \code{scan_menu()}.
}
\examples{
takeout(menu_crd(n = 50, t = 5))
# if you omit the design parameters then it will use the default
# (which may be random)
takeout(menu_crd())
# if you don't give any short names then it will generate a random one
takeout()
}
\seealso{
See \code{\link[=scan_menu]{scan_menu()}} for finding the short names of the
named experimental designs.
}
