% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{traits}
\alias{traits}
\title{Setting the traits of factors}
\usage{
traits(
  n = NULL,
  longname = NULL,
  desc = NULL,
  unit = NULL,
  labels = NULL,
  prefix = "",
  suffix = "",
  class = NULL,
  sep = edibble_labels_opt("sep"),
  include_leading_zero = edibble_labels_opt("leading_zero"),
  ...
)
}
\arguments{
\item{n}{The number of levels. This is ignored if \code{labels} is provided.}

\item{longname}{The long name of the variable or level.}

\item{desc}{The text description of the factor.}

\item{unit}{A string denoting the unit of measurement if applicable.}

\item{labels}{A string vector that denotes the short name of the levels.}

\item{prefix}{The prefix of the labels.}

\item{suffix}{The suffix of the labels.}

\item{include_leading_zero}{A logical value to indicate whether there should
be a leading zero added to level indexes. This is ignored if \code{labels} is provided.}

\item{...}{A name-value pair of attributes. The value must be a scalar and
attributed to the whole variable (not individual levels).
The values are added as attributes to the output object.}
}
\description{
This function is used to set characteristics of the factors.
}
