% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{utility-edibble-var}
\alias{utility-edibble-var}
\alias{as.character.edbl_var}
\alias{as.integer.edbl_var}
\alias{is_edibble_var}
\title{Utility functions for edibble variable}
\usage{
\method{as.character}{edbl_var}(x, ...)

\method{as.integer}{edbl_var}(x, ...)

is_edibble_var(x)
}
\arguments{
\item{x}{An \code{edbl_var} object.}

\item{...}{Ignored.}
}
\description{
The S3 methods for \code{edbl_var} objects have
the same expected output that of a factor.

Other functions are utility functions related to \code{edbl_var} object.
}
