% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_exists}
\alias{var_exists}
\title{Check if the design contains a particular variable label.}
\usage{
var_exists(design, label = NULL)
}
\arguments{
\item{design}{An edibble design.}

\item{label}{A character vector of the variable labels.}
}
\value{
A logical vector of the same size as \code{label} or a single logical value if class is provied.
}
\description{
Check if the design contains a particular variable label.
}
